% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absval_iwres_vs_pred.R
\name{absval_iwres_vs_pred}
\alias{absval_iwres_vs_pred}
\title{Absolute value of individual weighted residuals (IWRES) plotted against population predictions (PRED)}
\usage{
absval_iwres_vs_pred(xpdb, aes = NULL, group = "ID", type = "pls",
  by = NULL, layers = NULL, title = NULL, subtitle = NULL,
  log = FALSE, gg_theme = NULL, ...)
}
\arguments{
\item{xpdb}{an xpose database object.}

\item{aes}{ggxpose aesthetics (eg. \code{point_color}).}

\item{group}{grouping variable to be used.}

\item{type}{string setting the type of plot to be used points 'p',
line 'l' and smooth 's' or any combination of the 3.}

\item{by}{variable to be used for faceting.}

\item{layers}{a list of additional ggplot layers to be used.}

\item{title}{the main title of the plot. If NULL automated title will be generated.
Use FALSE to remove title and subtitle.}

\item{subtitle}{the plot subtitle. If NULL automated subtitle will be generated.
Use FALSE to remove subtitle.}

\item{log}{logical if TRUE axes will be logged.}

\item{gg_theme}{a ggplot2 theme to be used on this specific plot.}

\item{...}{any additional aesthetics.}
}
\description{
Plot of absolute value of individual weighted residuals (IWRES) plotted against population predictions (PRED).
}

