% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{add_facet_var}
\alias{add_facet_var}
\title{Add faceting variable}
\usage{
add_facet_var(facets, variable = "variable")
}
\arguments{
\item{facets}{User input faceting variable as a character or formula.}

\item{variable}{Default variable to be appended to variables in facets.}
}
\value{
facets as formula or character string.
}
\description{
Convenience function to add default faceting variable
}
\keyword{internal}
