% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{aes_c}
\alias{aes_c}
\title{Set new default value for ggplot2 aesthetics}
\usage{
aes_c(fun_aes, user_aes)
}
\arguments{
\item{fun_aes}{Default ggplot2 aesthetics mapping defined for the function}

\item{user_aes}{ggplot2 aesthetics mapping inputted by the user. These 
aesthetics will overwrite matching elements in `fun_aes`.}
}
\value{
ggplot2 aesthetics mapping
}
\description{
Set new default value for ggplot2 aesthetics
}
\keyword{internal}
