% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{aes_filter}
\alias{aes_filter}
\title{Convenience function to easily filter ggplot2 aesthetics}
\usage{
aes_filter(mapping, drop = NULL, keep_only = NULL)
}
\arguments{
\item{mapping}{ggplot2 aesthetics}

\item{drop}{names of ggplot2 aesthetics to be removed from the `mapping`. 
Alternative to the `keep_only` argument.}

\item{keep_only}{names of the only ggplot2 aesthetics to be kept in the 
`mapping`. Alternative to the `drop` argument.}
}
\value{
ggplot2 aesthetics mapping
}
\description{
Convenience function to easily filter ggplot2 aesthetics
}
\keyword{internal}
