% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{aes_rename}
\alias{aes_rename}
\title{Convenience function to easily rename ggplot2 aesthetics}
\usage{
aes_rename(mapping, from, to)
}
\arguments{
\item{mapping}{ggplot2 aesthetics}

\item{from}{names of ggplot2 aesthetics to be renamed. Note: should match the 
order of arg `to`.}

\item{to}{new names to be set for ggplot2 aesthetics. Note: should match the 
order of arg `from`.}
}
\value{
ggplot2 aesthetics mapping
}
\description{
Convenience function to easily rename ggplot2 aesthetics
}
\keyword{internal}
