% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{all_data_problem}
\alias{all_data_problem}
\title{Get all data problems}
\usage{
all_data_problem(xpdb)
}
\arguments{
\item{xpdb}{An xpose database object.}
}
\value{
The number of the all data problems as an integer 
vector.
}
\description{
Get all data problems
}
\keyword{internal}
