% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{all_file_problem}
\alias{all_file_problem}
\title{Get all file problems}
\usage{
all_file_problem(xpdb, ext)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{ext}{The file extension to be used.}
}
\value{
The number of the all file problems as an integer 
vector for the given file `ext`.
}
\description{
Get all file problems
}
\keyword{internal}
