% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.ctime}
\alias{as.ctime}
\title{Time converter}
\usage{
as.ctime(x)
}
\arguments{
\item{x}{A string ('character' or 'numeric') of time in seconds.}
}
\value{
A string of time in `hh:mm:ss`.
}
\description{
Convert seconds to hh:mm:ss
}
\keyword{internal}
