% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.nm.table.list}
\alias{as.nm.table.list}
\title{Convert an object to `nm_table_list` class}
\usage{
as.nm.table.list(x)
}
\arguments{
\item{x}{An object to be modified.}
}
\value{
x with `nm_table_list` class.
}
\description{
Adds `nm_table_list` attribute to an object
}
\keyword{internal}
