% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.table.list.nm}
\alias{as.table.list.nm}
\title{Convert an object to `table_list_nm` class}
\usage{
\method{as.table}{list.nm}(x)
}
\arguments{
\item{x}{An object to be modified.}
}
\value{
x with `table_list_nm` class.
}
\description{
Adds `table_list_nm` attribute to an object
}
\keyword{internal}
