% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.xpdb}
\alias{as.xpdb}
\title{Convert an object to `xpdb` class}
\usage{
as.xpdb(x)
}
\arguments{
\item{x}{An object to be modified.}
}
\value{
x with `xpdb` class.
}
\description{
Adds `xpdb` attribute to an object
}
\keyword{internal}
