% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.xpose.plot}
\alias{as.xpose.plot}
\title{Convert an object to `xpose_plot` class}
\usage{
as.xpose.plot(x)
}
\arguments{
\item{x}{An object to be modified.}
}
\value{
x with `xpose_plot` class.
}
\description{
Adds `xpose_plot` attribute to an object
}
\keyword{internal}
