% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.xpose.theme}
\alias{as.xpose.theme}
\title{Convert an object to `xpose_theme` class}
\usage{
as.xpose.theme(x)
}
\arguments{
\item{x}{An object to be modified.}
}
\value{
x with `xpose_theme` class.
}
\description{
Adds `xpose_theme` attribute to an object
}
\keyword{internal}
