% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_xpdb.R
\name{c.xpdb}
\alias{c.xpdb}
\title{Combine xpdb objects}
\usage{
\method{c}{xpdb}(...)
}
\arguments{
\item{...}{\code{\link{xpdb}} objects to be combined}
}
\description{
This function combines several \code{\link{xpdb}} object into a
  single one \code{\link{xpdb}}.
}
\details{
Only \code{\link{xpdb}} based on a single modeling software (e.g.
NONMEM, nlmixr) can be combined together. The themes (ggtheme, xptheme) and
options (i.e. quiet) will be taken from the first xpdb element.
}
\examples{
xpdb_combined <- c(xpdb_ex_pk, xpdb_ex_pk)
xpdb_combined

}
