% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{check_plot_type}
\alias{check_plot_type}
\title{Check plot `type`}
\usage{
check_plot_type(type, allowed)
}
\arguments{
\item{type}{A character string guiding the plot type.}

\item{allowed}{A character vector of all allowed types e.g. `c('p', 'l')`.}
}
\value{
Silent if proper user input else warns for unrecognized `type`.
}
\description{
Check plot `type`
}
\keyword{internal}
