% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{check_problem}
\alias{check_problem}
\title{Check .problem, .subprob and .method}
\usage{
check_problem(.problem, .subprob, .method)
}
\arguments{
\item{.problem}{The problem to be checked.}

\item{.subprob}{The subproblem to be checked.}

\item{.method}{The estimation method to be checked.}

\item{xpdb}{An xpose database object.}
}
\description{
Check .problem, .subprob and .method
}
\keyword{internal}
