% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_edits.R
\name{check_quo_vars}
\alias{check_quo_vars}
\title{Check quoted variables}
\usage{
check_quo_vars(xpdb, ..., .source, .problem)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{...}{Name-value pairs of expressions.}

\item{.source}{The source of the data in the xpdb. Can either be 'data' or an output 
file extension e.g. 'phi'.}

\item{.problem}{The problem from which the data will be modified}
}
\value{
Silent if checks are successful, returns errors otherwise.
}
\description{
Ensure that variables used in quos are present in the 
data and return informative error messages otherwise.
}
\keyword{internal}
