% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{check_scales}
\alias{check_scales}
\title{Check plot scales}
\usage{
check_scales(scale, log)
}
\arguments{
\item{scale}{The axis to be checked. Can be `x` or `y`.}

\item{log}{A string to guide which axis (e.g. `x`, `y` or `xy`) should be 
plotted on log scale.}
}
\value{
A string taking `continuous` or `log10` as value.
}
\description{
Check plot scales
}
\keyword{internal}
