% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_xpdb.R, R/xplot_helpers.R
\name{check_xpdb}
\alias{check_xpdb}
\alias{check_xpdb_source}
\alias{check_xpdb_changes}
\title{Check xpdb source files}
\usage{
check_xpdb_source(xpdb)

check_xpdb_changes(xpdb)

check_xpdb(xpdb, check = "data")
}
\arguments{
\item{xpdb}{An \code{xpdb} object generated with
\code{\link{create_nm_xpdb}}.}

\item{check}{The `xpdb` slot to be checked, can be `data`, `files`, 
`summary`, `special`, `code`, a file name, or `FALSE`. IF `FALSE` the slot 
test will be skipped.}

\item{xpdb}{An xpose database object.}
}
\value{
Silent if successful check, else returns error.
}
\description{
These functions check the status of the data in an xpdb.
  \code{\link{check_xpdb_source}} checks that the files used to build
  the xpdb have not been changed. \code{\link{check_xpdb_changes}} checks
  whether the xpdb content has been edited since its creation.
}
\examples{
# Check for changes in the source files
check_xpdb_source(xpdb_ex_pk)

# Check for changes in the xpdb
check_xpdb_changes(xpdb_ex_pk)

}
\keyword{internal}
