% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwres_vs_pred.R
\name{cwres_vs_pred}
\alias{cwres_vs_pred}
\title{Conditional residuals (CWRES) plotted against population predictions (PRED)}
\usage{
cwres_vs_pred(xpdb, aes = NULL, group = "ID", type = "pls", by = NULL,
  layers = NULL, title = NULL, subtitle = NULL, caption = NULL,
  log = FALSE, guides = TRUE, gg_theme = NULL, ...)
}
\arguments{
\item{xpdb}{an xpose database object.}

\item{aes}{ggxpose aesthetics (eg. \code{point_color}).}

\item{group}{grouping variable to be used.}

\item{type}{string setting the type of plot to be used points 'p',
line 'l' and smooth 's' or any combination of the 3.}

\item{by}{variable to be used for faceting.}

\item{layers}{a list of additional ggplot layers to be used.}

\item{title}{the main title of the plot. If NULL automated title will be generated.
Use FALSE to remove title and subtitle.}

\item{subtitle}{the plot subtitle. If NULL automated subtitle will be generated.
Use FALSE to remove subtitle.}

\item{caption}{page caption. If NULL automated caption will be generated.
Use FALSE to remove caption.}

\item{log}{logical if TRUE axes will be logged.}

\item{guides}{should the guides (eg. unity line) be displayed.}

\item{gg_theme}{a ggplot2 theme to be used on this specific plot.}

\item{...}{any additional aesthetics.}
}
\description{
Plot of conditional residuals (CWRES) plotted against population predictions (PRED).
}
\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

