% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{default_plot_problem}
\alias{default_plot_problem}
\title{Get the default data problem to be plotted}
\usage{
default_plot_problem(xpdb)
}
\arguments{
\item{xpdb}{An xpose database object.}
}
\value{
The last estimation problem when at least one estimation problem is present, 
else the last simulation problem as an integer.
}
\description{
Used by plot functions when the problem argument has not 
been supplied by the user.
}
\keyword{internal}
