% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distibution.R
\name{distrib_plot}
\alias{distrib_plot}
\alias{prm_distrib}
\alias{eta_distrib}
\alias{res_distrib}
\alias{cov_distrib}
\title{Distribution plots of ETA and parameters}
\usage{
prm_distrib(xpdb, mapping = NULL, type = "hr", facets = NULL,
  title = "Parameter distribution | @run",
  subtitle = "Based on @nind individuals", caption = "@dir", log = NULL,
  guides = FALSE, problem, quiet, ...)

eta_distrib(xpdb, mapping = NULL, type = "hr", facets = NULL,
  title = "Eta distribution | @run",
  subtitle = "Based on @nind individuals, Eta shrink: @etashk",
  caption = "@dir", log = NULL, guides = FALSE, problem, quiet, ...)

res_distrib(xpdb, mapping = NULL, res = "CWRES", type = "hr",
  facets = NULL, title = "@x distribution | @run",
  subtitle = "Based on @nobs observations", caption = "@dir", log = NULL,
  guides = FALSE, problem, quiet, ...)

cov_distrib(xpdb, mapping = NULL, type = "hr", facets = NULL,
  title = "Continuous covariates distribution | @run",
  subtitle = "Based on @nind individuals", caption = "@dir", log = NULL,
  guides = FALSE, problem, quiet, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{type}{String setting the type of plot to be used. Can be histogram 'h',
density 'd', rug 'r' or any combination of the three.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} 
or a formula to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 
'x', y' or 'xy'.}

\item{guides}{Should the guides (e.g. reference distribution) be displayed.}

\item{problem}{The $problem number to be used. By default returns 
the last estimation problem.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}

\item{res}{Only used for \code{res_distrib}. Defines the type of residual to be used. Default is "CWRES".}
}
\description{
Histograms and density plots of the ETA and parameter values.
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. histogram_fill = 'blue', rug_sides = 'b', etc.
\itemize{
  \item histogram: options to \code{geom_histogram}
  \item density: options to \code{geom_density}
  \item rug: options to \code{geom_rug}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Faceting}{

Every xpose plot function has built-in facetting functionalities. Faceting arguments 
are passed to the functions \link[ggforce]{facet_wrap_paginate} when the \code{facets} 
argument is a character string (e.g. \code{facets = c('OCC', 'SEX')}) or 
\link[ggforce]{facet_grid_paginate} when facets is a formula (e.g. \code{facets = OCC~SEX}).
All xpose plot functions accept all the arguments for the \link[ggforce]{facet_wrap_paginate} 
and \link[ggforce]{facet_grid_paginate} functions e.g. \code{dv_vs_ipred(xpdb_ex_pk, 
facets = OCC~SEX, ncol = 3, nrow = 3, page = 1, margins = TRUE, labeller = 'label_both')}.
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
# Histogram of parameters
prm_distrib(xpdb_ex_pk, type = 'h')

# Density plot of etas with a rug
eta_distrib(xpdb_ex_pk, type = 'dr')

# Histogram of residuals by occasion
res_distrib(xpdb_ex_pk, type = 'hr', res = c('IWRES', 'CWRES'))

# Density plot of continuous covariates
cov_distrib(xpdb_ex_pk, type = 'd')
}
\seealso{
\code{\link{xplot_distrib}}
}
