% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dv_vs_ipred.R
\name{dv_vs_ipred}
\alias{dv_vs_ipred}
\title{Observations (DV) plotted against individual predictions (IPRED)}
\usage{
dv_vs_ipred(xpdb, aes = NULL, group = "ID", type = "pls", by = NULL,
  layers = NULL, title = NULL, subtitle = NULL, caption = NULL,
  log = FALSE, guides = TRUE, gg_theme = NULL, ...)
}
\arguments{
\item{xpdb}{an xpose database object.}

\item{aes}{ggxpose aesthetics (eg. \code{point_color}).}

\item{group}{grouping variable to be used.}

\item{type}{string setting the type of plot to be used points 'p',
line 'l' and smooth 's' or any combination of the 3.}

\item{by}{variable to be used for faceting.}

\item{layers}{a list of additional ggplot layers to be used.}

\item{title}{the main title of the plot. If NULL automated title will be generated.
Use FALSE to remove title and subtitle.}

\item{subtitle}{the plot subtitle. If NULL automated subtitle will be generated.
Use FALSE to remove subtitle.}

\item{log}{logical if TRUE axes will be logged.}

\item{guides}{should the guides (eg. unity line) be displayed.}

\item{gg_theme}{a ggplot2 theme to be used on this specific plot.}

\item{...}{any additional aesthetics.}
}
\description{
Plot of observations (DV) vs individual predictions (IPRED).
}
