% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gof.R
\name{dv_vs_pred}
\alias{dv_vs_pred}
\alias{dv_vs_ipred}
\alias{dv_vs_pred}
\alias{dv_vs_cpred}
\title{Observations plotted against model predictions}
\usage{
dv_vs_ipred(xpdb, aes = NULL, group = "ID", type = "pls", facets = NULL,
  title = "DV vs. IPRED | @run",
  subtitle = "Ofv: @ofv, Eps shrink: @epsshk", caption = "@dir",
  log = NULL, guides = TRUE, ...)

dv_vs_pred(xpdb, aes = NULL, group = "ID", type = "pls", facets = NULL,
  title = "DV vs. PRED | @run", subtitle = "Ofv: @ofv", caption = "@dir",
  log = NULL, guides = TRUE, ...)

dv_vs_cpred(xpdb, aes = NULL, group = "ID", type = "pls", facets = NULL,
  title = "DV vs. CPRED | @run", subtitle = "Ofv: @ofv", caption = "@dir",
  log = NULL, guides = TRUE, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{aes}{The xpose aesthetics (eg. \code{point_color}).}

\item{group}{Grouping variable to be used for lines.}

\item{type}{String setting the type of plot to be used points 'p',
line 'l' and smooth 's' or any combination of the three.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} or a formula 
to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 'x', y' or 'xy'.}

\item{guides}{Enable guides display (e.g. unity line).}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}
}
\description{
Plot of observations (DV) vs population predictions (PRED), individual predictions (IPRED) 
or conditional population predictions (CPRED).
}
\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
dv_vs_pred(xpdb_ex_pk)

dv_vs_ipred(xpdb_ex_pk)

dv_vs_cpred(xpdb_ex_pk)

}
\seealso{
\code{\link{xplot_scatter}}
}
