% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distibution.R
\name{eta_distrib}
\alias{eta_distrib}
\title{Distribution plots of the ETA}
\usage{
eta_distrib(xpdb, mapping = NULL, type = "hr", facets = NULL,
  title = "@x distribution | @run",
  subtitle = "@nind individuals, @nobs observations", caption = "@dir",
  log = NULL, guides = FALSE, problem, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{type}{String setting the type of plot to be used points 'p',
line 'l', smooth 's' and text 't' or any combination of the four.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} 
or a formula to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 
'x', y' or 'xy'.}

\item{guides}{Enable guides display (e.g. unity line).}

\item{problem}{The $problem number to be used. By default returns 
the last estimation problem.}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}
}
\description{
Histograms or density plots plots ETA.
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. histogram_fill = 'blue', rug_sides = 'b', etc.
\itemize{
  \item histogram: options to \code{geom_histogram}
  \item density: options to \code{geom_density}
  \item panel: options to \code{facet_wrap} (facets is character) or \code{facet_grid} 
  (facets is a formula)
  \item rug: options to \code{geom_rug}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
# Histogram
eta_distrib(xpdb_ex_pk, type = 'h')

# Density plot with a rug
eta_distrib(xpdb_ex_pk, type = 'dr')

}
\seealso{
\code{\link{xplot_distrib}}
}
