% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{filter_xp_theme}
\alias{filter_xp_theme}
\title{Subset an `xp_theme`}
\usage{
filter_xp_theme(xp_theme, regex = NULL, action = "keep")
}
\arguments{
\item{xp_theme}{An xpose theme list.}

\item{regex}{A string used to find matching names in the `xp_theme`}

\item{action}{A string used to determine the action of on the xp_theme.
`keep` will subset the `xp_theme` to only the names matching the `regex`, 
while `drop` will drop the matches.}
}
\value{
A modified `xp_theme`
}
\description{
Subset an `xp_theme`
}
\keyword{internal}
