% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{get_aes}
\alias{get_aes}
\title{Extract variable from mapping argument}
\usage{
get_aes(mapping, aes)
}
\arguments{
\item{mapping}{A mapping object.}

\item{aes}{An aesthetic to be extracted}
}
\value{
The name of a variable.
}
\description{
Get the name of the variables used in a mapping argument, based
upon the coding standard used in the `xpose` package
}
\examples{
\dontrun{
library(ggplot2)
mapping <- aes(x = TIME, y = DV, point_color = MED1)
get_aes(mapping, 'x')
get_aes(mapping, 'point_colour') ## Beware "color" gets changed to "colour"
}
}
\keyword{internal}
