% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_access.R
\name{get_code}
\alias{get_code}
\title{Access model code}
\usage{
get_code(xpdb, parsed = TRUE, .problem = NULL)
}
\arguments{
\item{xpdb}{An \code{xpdb} object from which the model code will be
extracted.}

\item{parsed}{Should the parsed (\code{parsed = TRUE} (default)) or raw
(\code{parsed = FALSE}) model code be returned.}

\item{.problem}{The problem to be used (only used with \code{parsed = TRUE}).
Of note, problem 0 is attributed to general output (e.g. NM-TRAN
warnings in NONMEM). When ]code{NULL} (default) returns the entire code.}
}
\value{
A tibble of the parsed model (if \code{parsed = TRUE}) or a character
  vector of the raw model (if \code{parsed = FALSE}).
}
\description{
Access model code from an xpdb object.
}
\examples{
parsed_code <- get_code(xpdb_ex_pk)
head(parsed_code)

raw_code <- get_code(xpdb_ex_pk, parsed = FALSE)
head(raw_code)

}
\seealso{
\code{\link{create_xpdb}}, \code{\link{read_nm_model}}
}
