% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_access.R
\name{get_nm_prm_transformation_formulas}
\alias{get_nm_prm_transformation_formulas}
\title{Generate default transformation formulas for parameters}
\usage{
get_nm_prm_transformation_formulas(prm_names)
}
\arguments{
\item{prm_names}{Vector of parameter names as found in .ext}
}
\value{
List of formulas decribing the transformation
}
\description{
Generate default transformation formulas for parameters
}
\keyword{internal}
