% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_access.R
\name{get_special}
\alias{get_special}
\title{Access special model data}
\usage{
get_special(xpdb, .problem = NULL, quiet)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object from which the special data will be extracted.}

\item{.problem}{The problem to be used, by default returns the last one.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
A list.
}
\description{
Access special model data from an xpdb object.
}
\examples{
special <- get_summary(xpdb_ex_pk)
special

}
\seealso{
\code{\link{list_special}}, \code{\link{xpose_data}}
}
