% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_get.R
\name{get_vars}
\alias{get_vars}
\alias{get_var_types}
\alias{get_var_labels}
\alias{get_var_units}
\title{Get variable type, label or units}
\usage{
get_var_types(xpdb, ..., .problem = NULL)

get_var_labels(xpdb, ..., .problem = NULL)

get_var_units(xpdb, ..., .problem = NULL)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object.}

\item{.problem}{The problem number to which the edits will be applied.}

\item{variable}{A vector of data variables}
}
\value{
A character vector
}
\description{
Function designed to extract the type, label or unit associated
with variables
}
\examples{
# Get variable types
get_var_types(xpdb_ex_pk, .problem = 1, c('ID', 'MED1')

# Get variable labels
get_var_labels(xpdb_ex_pk, .problem = 1, c('ID', 'MED1')

# Get variable units
get_var_units(xpdb_ex_pk, .problem = 1, c('ID', 'MED1')

}
