% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_access.R
\name{grab_iter}
\alias{grab_iter}
\title{Grab parameter for a given iteration number}
\usage{
grab_iter(ext, iter)
}
\arguments{
\item{ext}{parameter vs. iteration table}

\item{iter}{iteration number}
}
\description{
Grab parameter for a given iteration number
}
\keyword{internal}
