% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individuals.R
\name{ind_plots}
\alias{ind_plots}
\title{Observations, individual predictions and population predictions plotted against 
the independent variable for every individual}
\usage{
ind_plots(xpdb, mapping = NULL, group = "variable", type = "lp",
  facets = NULL, title = "Individual plots | @run",
  subtitle = "Ofv: @ofv, Eps shrink: @epsshk", caption = "@dir",
  log = NULL, nrow = 3, ncol = 3, page = 1, problem, quiet, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{group}{Grouping variable to be used for lines.}

\item{type}{String setting the type of plot to be used points 'p',
line 'l', smooth 's' and text 't' or any combination of the four.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} 
or a formula to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 
'x', y' or 'xy'.}

\item{nrow}{Number of rows per page.}

\item{ncol}{Number of columns per page.}

\item{page}{Page to draw.}

\item{problem}{The $problem number to be used. By default returns 
the last estimation problem.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}
}
\description{
Observations (DV), individual predictions (IPRED) and population predictions (PRED) plotted against 
the independent variable for every individual
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. point_color = 'blue', smooth_method = 'lm', etc.
\itemize{
  \item point: options to \code{geom_point}
  \item line: options to \code{geom_line}
  \item guide: options to \code{geom_abline}
  \item panel: options to \code{facet_wrap} (facets is character) or \code{facet_grid} 
  (facets is a formula)
  \item smooth: options to \code{geom_smooth}
  \item text: options to \code{geom_text}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
# Basic example
ind_plots(xpdb_ex_pk)

# Example with mapping and facetting
ind_plots(xpdb_ex_pk, aes(x = TAD), 
          facets = OCC~ID, 
          panel_labeller = 'label_both')

}
\seealso{
\code{\link{xplot_scatter}}
}
