% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_edits.R
\name{irep}
\alias{irep}
\title{Add simulation counter}
\usage{
irep(x, quiet = FALSE)
}
\arguments{
\item{x}{The column to be used for computing simulation number, usually the ID column.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\description{
Add a column containing a simulation counter (irep). A new simulation is counted everytime
a value in x is lower than its previous value.
}
\examples{
xpdb_ex_pk_2 <- xpdb_ex_pk \%>\% 
 mutate(sim_id = irep(ID), .problem = 2)

}
