% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.formula}
\alias{is.formula}
\title{Test for formula class}
\usage{
is.formula(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `formula` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is a `formula` object
}
\keyword{internal}
