% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.nm.model}
\alias{is.nm.model}
\title{Test for nm_model class}
\usage{
is.nm.model(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `nm_model` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is a `nm_model` object
}
\keyword{internal}
