% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.table.list.nm}
\alias{is.table.list.nm}
\title{Test for table_list_nm class}
\usage{
is.table.list.nm(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `table_list_nm` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is a `table_list_nm` object
}
\keyword{internal}
