% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.xpose.plot}
\alias{is.xpose.plot}
\title{Test for xpose_plot class}
\usage{
is.xpose.plot(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `xpose_plot` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is a `xpose_plot` object
}
\keyword{internal}
