% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.xpose.theme}
\alias{is.xpose.theme}
\title{Test for xpose_theme class}
\usage{
is.xpose.theme(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `xpose_theme` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is an `xpose_theme` object
}
\keyword{internal}
