% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nm_input_record.R
\name{is_nm_reserved_label}
\alias{is_nm_reserved_label}
\title{Determine if a character string is a NONMEM reserved label}
\usage{
is_nm_reserved_label(x)
}
\arguments{
\item{x}{A character string (scalar or vector)}
}
\value{
A logical vector the same length as \code{x} indicating if the string
  is a NONMEM reserved label.
}
\description{
Determine if a character string is a NONMEM reserved label
}
