% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_df}
\alias{label_df}
\title{Add `mod_num` and `mod_file` columns}
\usage{
label_df(x, mod_num, mod_file)
}
\arguments{
\item{x}{A data frame.}

\item{mod_num}{The model rank.}

\item{mod_file}{A label for the model.}
}
\value{
The original dataframe preceded by the columns model and label.
}
\description{
Add the model rank (`mod_num`) and model file name (`mod_file`)
  to a data frame.
}
\keyword{internal}
