% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{last_data_problem}
\alias{last_data_problem}
\title{Get last data problem}
\usage{
last_data_problem(xpdb, simtab = FALSE)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{simtab}{Logical if `TRUE` the last simulation problem is returned. If `FALSE` 
the last estimation problem is returned.}
}
\value{
The last estimation problem when `simtab = FALSE`, 
else the last simulation problem as an integer.
}
\description{
Used by several internal functions when the problem argument has not 
been supplied by the user.
}
\keyword{internal}
