% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{last_file_problem}
\alias{last_file_problem}
\title{Get last file problem}
\usage{
last_file_problem(xpdb, ext)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{ext}{The file extension to be used.}
}
\value{
The number of the last file problem as an integer 
for the given file `ext`.
}
\description{
Get last file problem
}
\keyword{internal}
