% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nm_tab.R
\name{list_nm_tab}
\alias{list_nm_tab}
\title{List NONMEM output tables}
\usage{
list_nm_tab(mod_file = NULL, dir = NULL)
}
\arguments{
\item{mod_file}{A mod_file object generated with \code{\link{read_nm_model}}.}

\item{dir}{Location of the model files.}
}
\description{
List NONMEM output tables file names from a \code{mod_file} object.
}
\examples{
\dontrun{
file <- 'inst/extdata/run001.mod'
read_nm_model(file = file) \%>\% 
  list_nm_tab(dir = dirname(file)) \%>\%
  read_nm_tab()
  }
}
\seealso{
\code{\link{read_nm_model}}, \code{\link{read_nm_tab}}
}
