% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_list.R
\name{list_vars}
\alias{list_vars}
\title{List available variables}
\usage{
list_vars(xpdb, .problem = NULL)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object from which the model code will be extracted.}

\item{.problem}{The problem to be used, by lists all available problems.}
}
\value{
Prints the list of all available variables, and returns that list
  invisibly.  The name of the list is the problem number, the names of the
  elements of the sub-lists are the variable types, and the values of the
  sub-lists are the column names.
}
\description{
Function listing all available variables in an xpdb object.
}
\examples{
list_vars(xpdb_ex_pk)
}
\seealso{
\code{\link{set_var_types}}
}
