% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logLik.xpose_data}
\alias{logLik.xpose_data}
\title{Extract the log-likelihood from xpose_data}
\usage{
\method{logLik}{xpose_data}(object, ...)
}
\arguments{
\item{object}{The xpdb object}

\item{...}{passed to `get_summary()` and `get_prm()`.}
}
\value{
A logLik object including the likelihood, number of observations,
  number of individuals, and degrees of freedom (number of estimated
  parameters in the model).
}
\description{
Extract the log-likelihood from xpose_data
}
\examples{
logLik(xpdb_ex_pk)
AIC(xpdb_ex_pk)
BIC(xpdb_ex_pk)
nobs(logLik(xpdb_ex_pk))
}
