% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_minimization.R
\name{minimization_plots}
\alias{minimization_plots}
\alias{prm_vs_iteration}
\alias{grd_vs_iteration}
\title{Parameter value or gradient vs. iterations}
\usage{
prm_vs_iteration(xpdb, mapping = NULL, group = "variable", type = "pl",
  facets = NULL, title = "Parameter @y vs. @x | @run",
  subtitle = "Method: @method, minimization time: @runtime\\nTermination message: @term",
  caption = "@dir", log = NULL, guides = FALSE, problem, subprob, ...)

grd_vs_iteration(xpdb, mapping = NULL, group = "variable", type = "pl",
  facets = NULL, title = "Gradient @y vs. @x | @run",
  subtitle = "Method: @method, minimization time: @runtime\\nTermination message: @term",
  caption = "@dir", log = NULL, guides = FALSE, problem, subprob, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{group}{Grouping variable to be used for lines.}

\item{type}{String setting the type of plot to be used points 'p',
line 'l' and smooth 's' or any combination of the three.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} 
or a formula to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 
'x', y' or 'xy'.}

\item{guides}{Enable guides display (e.g. unity line).}

\item{problem}{The $problem number to be used. By default returns 
the last estimation problem.}

\item{subprob}{The sub-problem number to be used. By default returns 
the last sub-problem associated with the selected problem.}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}
}
\description{
Change of parameter value or gradient vs. iterations.
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. point_color = 'blue', smooth_method = 'lm', etc.
\itemize{
  \item point: options to \code{geom_point}
  \item line: options to \code{geom_line}
  \item guide: options to \code{geom_abline}
  \item panel: options to \code{facet_wrap} (facets is character) or \code{facet_grid} 
  (facets is a formula)
  \item smooth: options to \code{geom_smooth}
  \item text: options to \code{geom_text}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
prm_vs_iteration(xpdb_ex_pk)

grd_vs_iteration(xpdb_ex_pk)

}
\seealso{
\code{\link{xplot_scatter}}
}
