% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_pages.R
\name{multiple_pages}
\alias{multiple_pages}
\title{Allows \code{facet_wrap()} to work on multiple pages}
\usage{
multiple_pages(plot = NULL, by = NULL, ncol = 2, nrow = 2,
  scales = "fixed")
}
\arguments{
\item{plot}{a ggplot object}

\item{by}{variables to facet by}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{scales}{should scales be fixed ("fixed", the default), free ("free"),
or free in one dimension ("free_x", "free_y")}
}
\description{
Allows \code{ggplot2} panels to be plotted over multiple pages.
}
\examples{
\dontrun{
p <- ggplot(diamonds, aes(x = price, y = carat, color = cut)) +
geom_point(alpha = 0.5) +
labs(x = 'Price', y = 'Carat', title = 'Diamonds')

multiple_pages(plot = p, by = 'color', ncol = 2, nrow = 2)
}
}
\seealso{
\code{\link{facet_wrap}}
}
