% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_edits.R
\name{mutate.xpose_data}
\alias{mutate.xpose_data}
\title{Add new variables to an xpdb}
\usage{
\method{mutate}{xpose_data}(.data, ..., .problem, .source)
}
\arguments{
\item{.data}{An xpose database object.}

\item{...}{Name-value pairs of expressions. Use \code{NULL} to drop a variable.

These arguments are automatically quoted and evaluated in the 
context of the data frame. They support unquoting and splicing. 
See the dplyr vignette("programming") for an introduction to these concepts.}

\item{.problem}{The problem from which the data will be modified}

\item{.source}{The source of the data in the xpdb. Can either be 'data' or an output 
file extension e.g. 'phi'.}
}
\description{
\code{mutate()} adds new variables and preserves existing ones.
}
\examples{
xpdb_ex_pk \%>\% 
 mutate(TAD2 = TIME \%\% 40, .problem = 1) \%>\% 
 dv_vs_idv(aes(x = TAD2))
}
