% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{only_distinct}
\alias{only_distinct}
\title{Create functions for data deduplication}
\usage{
only_distinct(xpdb, .problem, facets, quiet)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{.problem}{The $problem number to be used.}

\item{facets}{The plot faceting variable. The `facets` variables along with the `id` column 
type will be as grouping factors during data deduplication process.}

\item{quiet}{Should messages be displayed to the console.}
}
\value{
A function
}
\description{
Create shortcut functions on the fly to remove duplicated records in data.
}
\keyword{internal}
