% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_geom.R
\name{parse_arg}
\alias{parse_arg}
\title{Parse arguments from xpose to ggplot2 format}
\usage{
parse_arg(x = NULL, name)
}
\arguments{
\item{x}{A list of xpose plot aesthetics.}

\item{name}{Name of the destination layer for the argument 
parsing (e.g. point).}
}
\value{
ggplot2 aesthetics for the layer defined in `name`.
}
\description{
Transform arguments from xpose (e.g. point_color) 
to ggplot2 (e.g. color) format.
}
\keyword{internal}
