% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_files.R
\name{parse_nm_files}
\alias{parse_nm_files}
\title{Parse NONMEM output files}
\usage{
parse_nm_files(file, name, raw, quiet)
}
\arguments{
\item{file}{The full file name (not used).}

\item{name}{The file name.}

\item{raw}{the raw data as a vector of strings.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
A tibble containing the parsed `data` along with `name`, `problem`, 
`subprob`, and `method`.
}
\description{
Function parsing NONMEM output files from their 
raw input.
}
\keyword{internal}
