% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nm_input_record.R
\name{parse_nm_input_record}
\alias{parse_nm_input_record}
\title{Extract the parameter name mapping from a NONMEM $INPUT block}
\usage{
parse_nm_input_record(code)
}
\arguments{
\item{code}{A source of data for a NONMEM $INPUT block (methods are available
for multiple sources)}
}
\value{
A named character vector mapping NONMEM reserved labels to input
  data.
}
\description{
Extract the parameter name mapping from a NONMEM $INPUT block
}
\examples{
parse_nm_input_record("FOO=ID")
parse_nm_input_record("FOO=ID BAR")
parse_nm_input_record(c("FOO=ID", "BAR"))
\dontrun{
# Read your model into xpdb
parse_nm_input_record(xpdb)
}
}
\seealso{
\code{\link{is_nm_reserved_label}}, \code{\link{read_nm_model}},
  \code{\link{read_nm_tables}}
}
