% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xpose_plot.R
\name{print.ggplot}
\alias{print.ggplot}
\title{Import print ggplot method}
\usage{
\method{print}{ggplot}(x, newpage = is.null(vp), vp = NULL, ...)
}
\description{
Import print ggplot method
}
\keyword{internal}
