% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xpdb.R
\name{print.xpdb_nm}
\alias{print.xpdb_nm}
\title{Print an xpdb object}
\usage{
\method{print}{xpdb_nm}(x, ...)
}
\arguments{
\item{x}{An \code{xpdb} object.}

\item{...}{Ignored in this function}
}
\description{
This function returns to the console a list of the files and
  options attached to an \code{\link{xpdb}} object.
}
\examples{
# Using the print function
print(xpdb_ex_pk)

# Or simply by writting the xpdb name
xpdb_ex_pk

}
