% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_print.R
\name{print.xpose_data}
\alias{print.xpose_data}
\title{Print an xpdb object}
\usage{
\method{print}{xpose_data}(xpdb)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object generated with \code{\link{xpose_data}}.}
}
\description{
This function will return a summary of the xpdb content to the console.
}
