% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.xpose_prm.R
\name{print.xpose_prm}
\alias{print.xpose_prm}
\title{Print an xpose_prm object}
\usage{
\method{print}{xpose_prm}(x, ...)
}
\arguments{
\item{x}{An \code{xpose_prm} object generated with \code{\link{get_prm}}.}

\item{...}{Ignored in this function}
}
\description{
This function returns to the console a list of the parameter estimates.
}
\examples{
# Create the xpose_prm object
prm <- get_prm(xpdb_ex_pk)

# Output using the print function
print(prm)

# Or simply by writting the xpdb name
prm

}
