% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{qq_plot}
\alias{qq_plot}
\alias{prm_qq}
\alias{eta_qq}
\alias{res_qq}
\alias{qq_plot}
\alias{cov_qq}
\title{QQ plots of ETA and residuals}
\usage{
prm_qq(xpdb, mapping = NULL, type = "p", facets = NULL,
  title = "QQ plot of parameters | @run",
  subtitle = "Based on @nind individuals", caption = "@dir", log = NULL,
  guide = TRUE, problem, quiet, ...)

eta_qq(xpdb, mapping = NULL, type = "p", facets = NULL,
  title = "QQ plot of etas | @run",
  subtitle = "Based on @nind individuals, Eta shrink: @etashk",
  caption = "@dir", log = NULL, guide = TRUE, problem, quiet, ...)

res_qq(xpdb, mapping = NULL, res = "CWRES", type = "p", facets = NULL,
  title = "QQ plot of @sample | @run",
  subtitle = "Based on @nobs observations", caption = "@dir", log = NULL,
  guide = TRUE, problem, quiet, ...)

cov_qq(xpdb, mapping = NULL, type = "p", facets = NULL,
  title = "QQ plot of continuous covariates | @run",
  subtitle = "Based on @nind individuals", caption = "@dir", log = NULL,
  guide = TRUE, problem, quiet, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{type}{String setting the type of plot. Can only be points 'p'.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} 
or a formula to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 
'x', y' or 'xy'.}

\item{guide}{Should the guide (e.g. reference line) be displayed.}

\item{problem}{The $problem number to be used. By default returns 
the last estimation problem.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}

\item{res}{Only used for \code{res_qq}. Defines the type of residual to be used. Default is "CWRES".}
}
\description{
QQ plots of the ETA and model residuals.
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. point_color = 'blue', etc.
\itemize{
  \item point: options to \code{geom_point}
  \item guide: options to \code{geom_abline}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Faceting}{

Every xpose plot function has built-in facetting functionalities. Faceting arguments 
are passed to the functions \link[ggforce]{facet_wrap_paginate} when the \code{facets} 
argument is a character string (e.g. \code{facets = c('OCC', 'SEX')}) or 
\link[ggforce]{facet_grid_paginate} when facets is a formula (e.g. \code{facets = OCC~SEX}).
All xpose plot functions accept all the arguments for the \link[ggforce]{facet_wrap_paginate} 
and \link[ggforce]{facet_grid_paginate} functions e.g. \code{dv_vs_ipred(xpdb_ex_pk, 
facets = OCC~SEX, ncol = 3, nrow = 3, page = 1, margins = TRUE, labeller = 'label_both')}.
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
# QQ plot of parameters
prm_qq(xpdb_ex_pk)

# QQ plot of eta
eta_qq(xpdb_ex_pk)

# QQ plot of residuals
res_qq(xpdb_ex_pk, res = c('IWRES', 'CWRES'))

# QQ plot of continuous covariates
cov_qq(xpdb_ex_pk)

}
\seealso{
\code{\link{xplot_distrib}}
}
