% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_file.R
\name{read_nm_file}
\alias{read_nm_file}
\title{NONMEM file parser}
\usage{
read_nm_file(filename = NULL, header = NULL, ext = NULL)
}
\arguments{
\item{filename}{Full file name, with either .ext, .phi. .grd, .cor or .cov extension.}

\item{header}{The name of the first header column (e.g 'NAME', 'ITERATION')}

\item{ext}{The file extenstion (e.g '.cov', '.ext')}
}
\value{
A \code{\link[dplyr]{tibble}} tables of containing the columns of each dataset. 
If mutiple headers are detected in the file then the last header is used. These columns are used to distinguish between problems: 
\itemize{
 \item{\strong{PROB}}{: the problem number}
 \item{\strong{SUBPROB}}{: The subproblem number.}
 \item{\strong{METHOD}}{: a character identifier named after the method used for each problem, i.e FOCE= First order conditional estimation.}
}
}
\description{
Parse NONMEM files in R format
}
\examples{
\dontrun{
nm_file <- read_nm_file(files = 'inst/extdata/run001.ext')
}
}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_tables}}
}
