% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_files.R
\name{read_nm_files}
\alias{read_nm_files}
\title{NONMEM file parser}
\usage{
read_nm_files(file = NULL, runno = NULL, dir = NULL, prefix = "run",
  suff = c(".ext", ".cor", ".cov", ".phi", ".grd"), verbose = TRUE)
}
\arguments{
\item{file}{Full file name, with either .ext, .phi. .grd, .cor or .cov extension. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{suff}.}

\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model file.}

\item{prefix}{Prefix of the model file name.}

\item{suff}{Extension of the model file.}

\item{verbose}{Logical, if \code{TRUE} messages are printed to the console.}
}
\value{
A list of \code{\link[dplyr]{tibble}} tables of class \code{} containing the following columns: 
\itemize{
 \item{\strong{ext}}{: The .ext file}
 \item{\strong{phi}}{: The .phi file}
 \item{\strong{grd}}{: The .grd file}
 \item{\strong{cov}}{: The .cov file.}
 \item{\strong{cor}}{: The .cor file.}
}
}
\description{
Parse NONMEM files in R format
}
\examples{
\dontrun{
nm_file <- read_nm_model(dir = '../models/pk/', runno = '001')
}
}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_tables}}
}
