% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_model.R
\name{read_nm_model}
\alias{read_nm_model}
\title{NONMEM model file parser}
\usage{
read_nm_model(runno = NULL, prefix = "run", ext = ".lst", file = NULL,
  dir = NULL, quiet = FALSE)
}
\arguments{
\item{runno}{Run number to be evaluated.}

\item{prefix}{Prefix of the model file name.}

\item{ext}{Extension of the model file. Should be one of ".lst", ".out", ".res", ".mod" or ".ctl".}

\item{file}{Full file name preferably a `.lst` file. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{ext}.}

\item{dir}{Location of the model file.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
A \code{\link[dplyr]{tibble}} of class \code{model} containing the following columns: 
\itemize{
 \item{\strong{problem}}{: a numeric identifier for the $PROBLEM associated with the code.}
 \item{\strong{level}}{: a unique numeric identifier to each subroutine block associated with the code.}
 \item{\strong{subroutine}}{: a character identifier named after the 3 first letters of the subroutine name e.g. "$THETA" and 
 "$TABLE" will become "the" and "tab" respectively. In addition all output from the .lst is labeled "lst", the general nonmem 
 output e.g. NM-TRAN messages are labeled "oth". With priors thp, tpv, omp, opd, sip, spd abbreviations are given to the THETAP, 
 THETAPV, OMEGAP, etc.}
 \item{\strong{code}}{: the code without comments or subroutine names e.g. "$THETA 0.5 ; TVCL" will return 0.5.}
 \item{\strong{comment}}{: the last comment of a record e.g. "0.5 ; Clearance (L/h) ; TVCL" will return "TVCL".}
}
}
\description{
Parse NONMEM model files in R format
}
\section{File path generation}{

The rules for model file names generation are as follow:
\itemize{
  \item with \code{runno}: the full path is generated as \code{<dir>/<prefix><runno>.<ext>} e.g. with \code{dir = 'model/pk'}, \code{prefix = 'run'}, \code{runno = '001'}, 
  \code{ext = '.lst'} the resulting path would be \code{model/pk/run001.lst}
  \item with \code{file}: the full path is generated as \code{<dir>/<file>} e.g. with \code{dir = 'model/pk'}, \code{file = 'run001.lst'} the resulting path
  would also be \code{model/pk/run001.lst}. Note: in this case the file extension should be provided as part of the `file` argument.
  }
}

\examples{
\dontrun{
# Using the `file` argument to import a model file:
nm_model <- read_nm_model(file = 'run001.lst', dir = 'models')

# Using the `runno` argument to import a model file:
nm_model <- read_nm_model(runno = '001', ext = '.lst', dir = 'models')
}

}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_tables}}
}
