% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{slice}
\alias{reexports}
\alias{distinct}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{summarize}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{summarise}}}

  \item{purrr}{\code{\link[purrr]{\%>\%}}}
}}

