% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{reorder_factors}
\alias{reorder_factors}
\title{Reorder factors by numerical order}
\usage{
reorder_factors(prefix, suffix = NULL)
}
\arguments{
\item{prefix}{A prefix to be added in front of the factor digits.}

\item{suffix}{A suffix to be added after the factor digits.}
}
\value{
A modified tibble
}
\description{
Will for example convert `ETA(1)` to 1 create factors then generate labels
by wrapping the digits with prefix and suffix.
}
\keyword{internal}
