% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_set.R
\name{set_vars_type}
\alias{set_vars_type}
\title{Set variable type}
\usage{
set_vars_type(xpdb, problem = NULL, ..., auto_factor = TRUE, quiet)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object.}

\item{problem}{The problem number to which the edits will be applied.}

\item{...}{Specifications of the edits to be made to the xpdb index. Edits are made as 
type and variable pairs e.g. idv = 'TAD' will assign TAD to the type idv (independent variable).}

\item{auto_factor}{If \code{TRUE} new columns assigned to the type 'catcov' will be converted to
factor.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
An xpose_data object
}
\description{
Function designed to change the type associated with variables.
}
\section{Recognized variable types}{

\itemize{
  \item amt: Dose amount
  \item catcov: Categorical covariates
  \item contcov: Continuous covariates
  \item dv: Dependent variable
  \item dvid: DV identifier
  \item eta: Eta
  \item evid: Event identifier
  \item id: Subject identifier
  \item idv: Independent variable
  \item ipred: Model individual predictions
  \item mdv: Missing dependent variable
  \item na: Not attributed
  \item occ: Occasion flag
  \item param: Model parameter
  \item pred: Model typical predictions
  \item res: Residuals
 }
}

\examples{
xpdb_ex_pk <- set_vars_type(xpdb_ex_pk, problem = 1, idv = 'TAD')
}
\seealso{
\code{\link{list_vars}}
}
