% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{software}
\alias{software}
\title{Get software from the xpdb}
\usage{
software(xpdb)
}
\arguments{
\item{xpdb}{An xpose database object.}
}
\value{
The name of a modeling tool.
}
\description{
Get the name of the modeling softwre on which the xpdb was created
}
\keyword{internal}
