% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_data.R
\name{[[<-.xpose_data}
\alias{[[<-.xpose_data}
\title{Allow assignment into xpose_data without conversion to class uneval}
\usage{
\method{[[}{xpose_data}(x, i) <- value
}
\description{
Allow assignment into xpose_data without conversion to class uneval
}
