% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_edits.R
\name{summarize_xpdb}
\alias{summarize_xpdb}
\alias{group_by.xpose_data}
\alias{summarize_xpdb}
\alias{ungroup.xpose_data}
\alias{summarize_xpdb}
\alias{summarize.xpose_data}
\alias{summarize_xpdb}
\alias{summarise.xpose_data}
\title{Group/ungroup and summarize variables in an xpdb}
\usage{
\method{group_by}{xpose_data}(.data, ..., .problem, .source)

\method{ungroup}{xpose_data}(.data, ..., .problem, .source)

\method{summarize}{xpose_data}(.data, ..., .problem, .source)

\method{summarise}{xpose_data}(.data, ..., .problem, .source)
}
\arguments{
\item{.data}{An xpose database object.}

\item{...}{Logical predicates defined in terms of the variables in .data. 
Multiple conditions are combined with &. Only rows where the condition evaluates 
to \code{TRUE} are kept. 
These arguments are automatically quoted and evaluated in the 
context of the data frame. They support unquoting and splicing. 
See the dplyr vignette("programming") for an introduction to these concepts.}

\item{.problem}{The problem from which the data will be modified}

\item{.source}{The source of the data in the xpdb. Can either be 'data' or an output 
file extension e.g. 'phi'.}
}
\description{
\code{group_by()} takes an existing table and converts it into a 
grouped table where operations are performed "by group". \code{ungroup()} removes grouping.
\code{summarize()} reduces multiple values down to a single value.
}
\examples{
# Create a distribution plot of Cmax
xpdb_ex_pk \%>\% 
 group_by(ID, SEX, .problem = 1) \%>\% 
 summarize(CMAX = max(DV), .problem = 1) \%>\% 
 ungroup(.problem = 1) \%>\% 
 xplot_distrib(aes(x = CMAX, density_fill = SEX), type = 'dr')
 
}
