% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_xpdb.R
\name{summary.xpdb_nm}
\alias{summary.xpdb_nm}
\title{Summarizing xpdb}
\usage{
\method{summary}{xpdb_nm}(object, .problem = NULL, ...)
}
\arguments{
\item{object}{An \code{xpdb} object.}

\item{.problem}{The problem to be used, by default returns the last one for
each label.}

\item{...}{Ignored in this function}
}
\description{
This function returns a summary of an \code{\link{xpdb}} to the
  console.
}
\examples{
summary(xpdb_ex_pk)

}
