% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_xpose_data.R
\name{summary.xpose_data}
\alias{summary.xpose_data}
\title{Summarizing xpose_data}
\usage{
\method{summary}{xpose_data}(object, .problem = NULL, ...)
}
\arguments{
\item{object}{An \code{xpose_data} object generated with \code{\link{xpose_data}}.}

\item{.problem}{The problem to be used, by default returns the last one for each label.}

\item{...}{Ignored in this function}
}
\description{
This function returns a summary of an \code{\link{xpose_data}} 
to the console.
}
\examples{
summary(xpdb_ex_pk)

}
