% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_titles.R
\name{template_titles}
\alias{template_titles}
\title{Template titles}
\examples{
# Defined when creating a plot
dv_vs_ipred(xpdb_ex_pk, 
            title = 'DV vs. IPRED',
            subtitle = '@ofv, @nind subjects, @nobs obs.',
            caption = '@run, @descr')
            
# Any label can be modified later on
dv_vs_ipred(xpdb_ex_pk, aes(point_color = SEX, 
                            line_color = SEX)) + 
 labs(title = 'This runs is: @descr', 
      color = 'Color scale for @run',
      x = 'IPRED for @nind subjects',
      subtitle = NULL)

}
\seealso{
\link{xpose_save}
}
