% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bw2.R
\name{theme_bw2}
\alias{theme_bw2}
\title{ggplot2 theme}
\usage{
theme_bw2(base_size = 11, base_family = "", legend_position = "right")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{legend_position}{the position of legends ("none", "left", "right", "bottom",
"top", or two-element numeric vector)}
}
\description{
Black and whithe theme intended to make ggplot2 more readable when used
in presentation or papers. Inspired by a theme from Gunnar Yngman.
}
\examples{
\dontrun{
ggplot(diamonds, aes(x = price, y = carat, color = cut)) +
 geom_point(alpha = 0.5) +
 theme_bw2(legend_position = 'top')
}
}
