% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_readable.R
\name{theme_readable}
\alias{theme_readable}
\title{ggplot2 theme}
\usage{
theme_readable(base_size = 11, legend.position = "right")
}
\arguments{
\item{base_size}{base font size}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom",
"top", or two-element numeric vector)}
}
\description{
Theme intended to make ggplot2 more readable when used
in presentation or papers. Background and grid lines were dimed to focus
the attention on the data itself.
}
\examples{
\dontrun{
ggplot(diamonds, aes(x = price, y = carat, color = cut)) +
 geom_point(alpha = 0.5) +
 theme_readable(legend.position = 'top')
}
}
