% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xp_default.R
\name{theme_xp_default}
\alias{theme_xp_default}
\title{Default ggxpose theme}
\usage{
theme_xp_default()
}
\description{
The default ggxpose theme.
}
\examples{
# Standard ggxpose theme
\dontrun{
 xpose_data(dir = 'inst/models/', runno = '001') \%>\%
   xpose_theme(gg_theme = theme_readable(),
               xp_theme = theme_xp_default()) \%>\%
   dv_vs_ipred()
}
}
