% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xp_xpose4.R
\name{theme_xp_xpose4}
\alias{theme_xp_xpose4}
\title{Xpose4 theme for ggxpose}
\usage{
theme_xp_xpose4()
}
\description{
An xp_theme that makes ggxpose look like xpose4.
}
\examples{
# Make ggxpose look like xpose4
\dontrun{
 xpose_data(dir = 'inst/models/', runno = '001') \%>\%
   xpose_theme(gg_theme = theme_bw(),
               xp_theme = theme_xp_xpose4()) \%>\%
   dv_vs_ipred()
}
}
