% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_geom.R
\name{update_args}
\alias{update_args}
\title{Update `xpose_geoms` arguments}
\usage{
update_args(thm_arg, name, ...)
}
\arguments{
\item{thm_arg}{A subset of `xp_theme` used to test defaults.}

\item{name}{Name of the destination layer for the argument 
parsing (e.g. point).}

\item{...}{User arguments.}
}
\value{
A list of arguments for the layer defined in `name`.
}
\description{
Combine the arguments from the user and the `xp_theme`.
}
\keyword{internal}
