% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_labs.R
\name{update_xpose_labels}
\alias{update_xpose_labels}
\title{Update axis/legend labels for xpose plots}
\usage{
update_xpose_labels(p, labels)
}
\arguments{
\item{p}{xpose plot object to modify}

\item{labels}{named list of new labels
#@export}
}
\description{
Function based on \link[ggplot2]{update_labels} enabling the 
use of \code{\link{template_titles}} using the ggplot2 \code{+} operator.
}
