% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_data.R
\name{vpc_data}
\alias{vpc_data}
\title{Import and compute data for visual predictive plots}
\usage{
vpc_data(xpdb = NULL, psn_folder = NULL, bins = "jenks",
  n_bins = "auto", bin_mid = "mean", obs_cols = NULL, sim_cols = NULL,
  stratify = NULL, pred_corr = FALSE, pred_corr_lower_bnd = 0,
  pi = c(0.05, 0.95), ci = c(0.05, 0.95), uloq = NULL, lloq = NULL,
  smooth = TRUE, obs_problem, sim_problem, quiet, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{psn_folder}{Optional, specify a PsN-generated VPC-folder.}

\item{bins}{Binning method, can be one of 'density', 'time', 'data', 'none', or one of the approaches 
available in \code{classInterval()} such as 'jenks' (default), 'pretty', or a numeric vector specifying 
the bin separators.}

\item{n_bins}{When using the 'auto' binning method, what number of bins to aim for.}

\item{bin_mid}{Specify how to is the mid bin value calculated, can be either 'mean' for the mean of all 
timepoints (default) or 'middle' to use the average of the bin boundaries.}

\item{obs_cols}{Observation dataset column names (list elements: 'dv', 'idv', 'id', 'pred')}

\item{sim_cols}{Simulation dataset column names (list elements: 'dv', 'idv', 'id', 'pred')}

\item{stratify}{Character vector of stratification variables. Only 1 or 2 stratification variables can be supplied.}

\item{pred_corr}{Logical, should a prediction correction (pcVPC) of the data be used.}

\item{pred_corr_lower_bnd}{Lower bound for the prediction-correction.}

\item{pi}{Simulated prediction interval to plot. Default is c(0.05, 0.95).}

\item{ci}{Confidence interval around the percentiles to plot. Default is (0.05, 0.95)}

\item{uloq}{Number or NULL indicating upper limit of quantification. Default is NULL.}

\item{lloq}{Number or NULL indicating lower limit of quantification. Default is NULL.}

\item{smooth}{Should the bins be smoothed (connect bin midpoints, default) or shown as rectangular boxes.}

\item{obs_problem}{Alternative to the option `psn_folder`. The $problem number to be used for observations.
By default returns the last estimation problem.}

\item{sim_problem}{Alternative to the option `psn_folder`. The $problem number to be used for simulations.
By default returns the last simulation problem.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{any additional aesthetics.}

\item{obs_problem}{Alternative to the option `psn_folder`. The $problem number to be used for observations.
By default returns the last estimation problem.}

\item{sim_problem}{Alternative to the option `psn_folder`. The $problem number to be used for simulations.
By default returns the last simulation problem.}
}
\description{
Function designed to import and compute data for VPC. This function is a modified version of the vpc::vpc() function
developped by Ron Keizer \url{https://github.com/ronkeizer/vpc}.
}
