% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc.R
\name{vpc_opt_set}
\alias{vpc_opt_set}
\title{Create options for data import}
\usage{
vpc_opt_set(bins = "jenks", n_bins = "auto", bin_mid = "mean",
  obs_cols = NULL, sim_cols = NULL, pred_corr = FALSE,
  pred_corr_lower_bnd = 0, ci = c(0.05, 0.95), pi = c(0.05, 0.95),
  lloq = NULL, uloq = NULL)
}
\arguments{
\item{bins}{Binning method, can be one of 'density', 'time', 'data', 'none', or one of the approaches 
available in \code{classInterval()} such as 'jenks' (default), 'pretty', or a numeric vector specifying 
the bin separators.}

\item{n_bins}{When using the 'auto' binning method, what number of bins to aim for.}

\item{bin_mid}{Specify how to is the mid bin value calculated, can be either 'mean' for the mean of all 
timepoints (default) or 'middle' to use the average of the bin boundaries.}

\item{obs_cols}{Observation dataset column names (list elements: 'dv', 'idv', 'id', 'pred')}

\item{sim_cols}{Simulation dataset column names (list elements: 'dv', 'idv', 'id', 'pred')}

\item{pred_corr}{Logical, should a prediction correction (pcVPC) of the data be used.}

\item{pred_corr_lower_bnd}{Lower bound for the prediction-correction.}

\item{ci}{Confidence interval around the percentiles to plot. Default is (0.05, 0.95)}

\item{pi}{Simulated prediction interval to plot. Default is c(0.05, 0.95).}

\item{lloq}{Number or NULL indicating lower limit of quantification. Default is NULL.}

\item{uloq}{Number or NULL indicating upper limit of quantification. Default is NULL.}
}
\description{
Provide a list of options to the general plotting functions such as 
\code{xplot_scatter} in order to create appropriate data input for ggplot2.
}
\examples{
vpc_opt_set()

}
\seealso{
\code{{vpc}}
}
