% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_labs.R
\name{+.gg}
\alias{+.gg}
\title{Add components to a ggplot2 or an xpose plot}
\usage{
\method{+}{gg}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{\link{ggplot}}, xpose_plot,
or a \code{\link{theme}}.}

\item{e2}{A plot component}
}
\description{
Function based on \link[ggplot2]{+.gg} enabling the 
use of \code{\link{template_titles}} using the ggplot2 \code{+} operator.
}
\seealso{
\link[ggplot2]{ggplot2}, \link[ggplot2]{+.gg}
}
