% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_geom.R
\name{xp_map}
\alias{xp_map}
\title{ggplot2 layer call}
\usage{
xp_map(arg, mapping, ggfun)
}
\arguments{
\item{arg}{A list of arguments for the target layer.}

\item{mapping}{ggplot2 aesthetics for the target layer.}

\item{ggfun}{Name of the ggplot2 layer function to be called.}
}
\value{
Output of the `ggfun` call.
}
\description{
Function calling the ggplot2 layer function 
with its parsed arguments.
}
\keyword{internal}
