% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xp_theme_default.R
\name{xp_theme_default}
\alias{xp_theme_default}
\title{Default ggxpose theme}
\description{
The default ggxpose theme.
}
\examples{
# Standard ggxpose theme
\dontrun{
 xpose_data(dir = 'inst/models/', runno = '001') \%>\%
   xpose_theme(gg_theme = theme_readable(),
               xp_theme = xp_theme_default) \%>\%
   dv_vs_ipred()
}
}
