% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_xp.R
\name{xp_themes}
\alias{xp_themes}
\alias{theme_xp_default}
\alias{theme_xp_xpose4}
\title{A set of xpose themes}
\usage{
theme_xp_default()

theme_xp_xpose4()
}
\description{
xpose themes are used to consistently apply a set of preference for the plot geoms 
(e.g. color scales, point size, etc.) whereas ggplot2 theme focus on the plot background, axes, 
titles etc.

\itemize{
\item \code{theme_xp_default}: The default xp_theme in xpose
\item \code{theme_xp_xpose4}: An xp_theme that makes xpose look like xpose4.
}
}
\examples{
# With the xp_theme theme_xp_default() (default)
dv_vs_ipred(xpdb_ex_pk, facets = 'SEX')

# With the xp_theme theme_xp_xpose4()
xpdb_ex_pk \%>\%
  update_themes(xp_theme = theme_xp_xpose4()) \%>\%
  dv_vs_ipred(facets = 'SEX')
  
}
