% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{xpdb_ex_pk}
\alias{xpdb_ex_pk}
\title{\code{xpose_data} examples}
\format{An \code{xpose_data} object}
\source{
\href{http://link.springer.com/article/10.1023/A:1020561807903}{Moxonidine model}: 
Assumption Testing in Population Pharmacokinetic Models: Illustrated with an Analysis of 
Moxonidine Data from Congestive Heart Failure Patients. Journal of Pharmacokinetics and 
Biopharmaceutics. 26(2):207–246 (1998).
}
\description{
Moxonidine \code{xpose_data} example.
}
\examples{
print(xpdb_ex_pk)

}
