% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_edits.R
\name{xpdb_index_update}
\alias{xpdb_index_update}
\title{Update data index}
\usage{
xpdb_index_update(xpdb, .problem)
}
\arguments{
\item{xpdb}{An xpose database object from which the index will be updated.}

\item{.problem}{The problem from which the index will be updated.}
}
\description{
Function dedicated to update the data index after modifications.
}
\keyword{internal}
