% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_distrib.R
\name{xplot_distrib}
\alias{xplot_distrib}
\title{Default xpose distribution plot function}
\usage{
xplot_distrib(xpdb, mapping = NULL, type = "hr", guides = FALSE,
  xscale = "continuous", yscale = "continuous", title = NULL,
  subtitle = NULL, caption = NULL, plot_name = "density_plot", gg_theme,
  xp_theme, data_opt, quiet, ...)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object generated with \code{\link{xpose_data}}.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{type}{String setting the type of plot to be used histogram 'h',
density 'd', rug 'r' the three.}

\item{guides}{Should the guides (e.g. reference distribution) be displayed.}

\item{xscale}{Scale type for x axis (eg. 'continuous', 'discrete', 'log10').}

\item{yscale}{Scale type for y axis (eg. 'continuous', 'discrete', 'log10').}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{plot_name}{Name to be used by \code{xpose_save()} when saving the plot.}

\item{gg_theme}{A ggplot2 theme object (eg. \code{\link[ggplot2]{theme_classic}}).}

\item{xp_theme}{An xpose theme or vector of modifications to the xpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')}).}

\item{data_opt}{A list of options in order to create appropriate data input for 
ggplot2. For more information see \code{\link{data_opt_set}}.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Any additional aesthetics.}
}
\description{
Manually generate distribution plots from an xpdb object.
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. histogram_fill = 'blue', rug_sides = 'b', etc.
\itemize{
  \item histogram: options to \code{geom_histogram}
  \item density: options to \code{geom_density}
  \item panel: options to \code{facet_wrap} (facets is character) or \code{facet_grid} 
  (facets is a formula)
  \item rug: options to \code{geom_rug}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
# A simple histogram
xplot_distrib(xpdb_ex_pk, aes(x = WT))

# A simple density plot
xplot_distrib(xpdb_ex_pk, aes(x = CWRES))

}
\seealso{
\code{\link{xplot_scatter}}
}
