% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_qq.R
\name{xplot_qq}
\alias{xplot_qq}
\title{Default xpose QQ plot function}
\usage{
xplot_qq(xpdb, mapping = NULL, type = "p", guides = FALSE,
  xscale = "continuous", yscale = "continuous", title = NULL,
  subtitle = NULL, caption = NULL, plot_name = "qq_plot", gg_theme,
  xp_theme, opt, quiet, ...)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object generated with \code{\link{xpose_data}}.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{type}{String setting the type of plot to be used. Can only be points 'p'.}

\item{guides}{Should the guides (e.g. reference line) be displayed.}

\item{xscale}{Scale type for x axis (eg. 'continuous', 'discrete', 'log10').}

\item{yscale}{Scale type for y axis (eg. 'continuous', 'discrete', 'log10').}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{plot_name}{Name to be used by \code{xpose_save()} when saving the plot.}

\item{gg_theme}{A ggplot2 theme object (eg. \code{\link[ggplot2]{theme_classic}}).}

\item{xp_theme}{An xpose theme or vector of modifications to the xpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')}).}

\item{opt}{A list of options in order to create appropriate data input for 
ggplot2. For more information see \code{\link{data_opt}}.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Any additional aesthetics.}
}
\description{
Manually generate QQ plots from an xpdb object.
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. point_color = 'blue', etc.
\itemize{
  \item point: options to \code{geom_point}
  \item guide: options to \code{geom_abline}
  \item panel: options to \code{facet_wrap} (facets is character) or \code{facet_grid} 
  (facets is a formula)
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
xplot_qq(xpdb_ex_pk, aes(sample = CWRES), guides = TRUE)

}
\seealso{
\code{\link{xplot_scatter}} \code{\link{xplot_distrib}}
}
