% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.R
\docType{package}
\name{xpose-package}
\alias{xpose}
\alias{xpose-package}
\title{xpose: graphical diagnostics for pharmacometric models}
\description{
xpose was designed as an interface between pharmacometrics software outputs and modelelers. 
This package aims to facilitates dataset checkout, exploration and visualization, model diagnostics, candidate covariate 
identification and model comparison.

To learn more about xpose, visit our website: \url{https://guiastrennec.github.io/xpose/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/guiastrennec/xpose}
  \item Report bugs at \url{https://github.com/guiastrennec/xpose/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Guiastrennec \email{guiastrennec@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Anna Olofsson
  \item Andrew C. Hooker \email{andrew.hooker@farmbio.uu.se}
}

Other contributors:
\itemize{
  \item Sebastian Ueckert [contributor]
  \item Kajsa Harling [contributor]
  \item Mike K. Smith [contributor]
  \item Elodie Plan [contributor]
  \item Pharmetheus [contributor]
  \item Pfizer [contributor]
}

}
