% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_geom.R
\name{xpose_panels}
\alias{xpose_panels}
\title{Generic panel function for `xpose_plots`}
\usage{
xpose_panels(xp_theme, extra_args)
}
\arguments{
\item{xp_theme}{An `xpose_theme` object.}

\item{extra_args}{User arguments to be passed to the 
faceting functions.}
}
\value{
Output a `facet` layer. Layer will be `facet_wrap_paginate` if 
`facets` is a string, and `facet_grid_paginate` if `facets` is a formula.
}
\description{
Convenience wrapper around ggforce faceting functions.
}
\keyword{internal}
