% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_plot_default.R
\name{xpose_plot_default}
\alias{xpose_plot_default}
\title{Default xpose plot}
\usage{
xpose_plot_default(xpdb, vars = NULL, aes = NULL, group = "ID",
  type = "pls", layers = NULL, title = NULL, subtitle = NULL,
  caption = NULL, guides = TRUE, xscale = "continuous",
  yscale = "continuous", gg_theme = NULL,
  plot_name = "xpose_plot_default", ...)
}
\arguments{
\item{xpdb}{an xpose database object.}

\item{vars}{the variable mapping from an aes().}

\item{aes}{ggxpose aesthetics (eg. \code{point_color}).}

\item{group}{grouping variable to be used.}

\item{type}{string setting the type of plot to be used points 'p',
line 'l', smooth 's' and text 't' or any combination of the 4.}

\item{layers}{a list of additional ggplot layers to be used.}

\item{title}{the main title of the plot. If NULL automated title will be generated.
Use FALSE to remove title and subtitle.}

\item{subtitle}{the plot subtitle. If NULL automated subtitle will be generated.
Use FALSE to remove subtitle.}

\item{caption}{page caption. If NULL automated caption will be generated.
Use FALSE to remove caption.}

\item{guides}{should the guides (eg. unity line) be displayed.}

\item{xscale}{scale type for x axis (eg. 'continuous', 'discrete', 'log10').}

\item{yscale}{scale type for y axis (eg. 'continuous', 'discrete', 'log10').}

\item{gg_theme}{a ggplot2 theme to be used on this specific plot.}

\item{plot_name}{name that will be used by \code{xpose_save()} to save the plot.}

\item{...}{any additional aesthetics.}
}
\value{
An \code{xpose_plot}
}
\description{
Manually generate plotting functions.
}
\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
\dontrun{
xpose_plot_default(xpdb, vars = aes(x = IPRED, y = DV))
}
}
