% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_save.R
\name{xpose_save}
\alias{xpose_save}
\title{Save ggxpose plot}
\usage{
xpose_save(plot = NULL, dir = NULL, name = NULL, format = "pdf",
  width = 12, height = 9, ...)
}
\arguments{
\item{plot}{a ggxpose plot object.}

\item{dir}{an optional path to a specific directory.}

\item{name}{an optional name to be given to the file. If left NULL the run
number and plotting function will be used instead.}

\item{format}{the file format to use: 'pdg', 'jpg' or 'png',}

\item{width}{the page width in inches}

\item{height}{the page height in inches}

\item{...}{additional options to be passed to \code{jpg()} or \code{png()} functions.}
}
\description{
Automatically save ggxpose plots to pdf, jpg or png.

If no arguments are provided, the function will automatically name
and save a plot after its run number and the plotting function name
by using the metadata attached to the plot.
}
\examples{
\dontrun{
dv_vs_ipred(xpdb) \%>\% xpose_save()
}
}
