% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_theme.R
\name{xpose_theme}
\alias{xpose_theme}
\title{Create ggxpose theme}
\usage{
xpose_theme(xpdb = NULL, gg_theme = NULL, xp_theme = NULL)
}
\arguments{
\item{xpdb}{a ggxpose database object}

\item{gg_theme}{a ggplot2 complete theme object (eg. \code{theme_classic()})}

\item{xp_theme}{a ggxpose theme or vector of modifications of the ggxpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')})}
}
\value{
An xpdb object
}
\description{
Create a ggxpose theme.
This function will update the theme of an xpdb object.
All plots generated with this xpdb will automatically use the
defined ggxpose (\code{xp_theme}) and ggplot2 (\code{gg_theme}) themes.
}
\examples{
\dontrun{
xpdb <- xpose_theme(xpdb     = xpdb,
                    gg_theme = theme_bw(),
                    xp_theme = c(point_color = 'blue',
                                 line_color  = 'blue'))
}
}

