% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_helpers.R
\name{check_xpdb}
\alias{check_xpdb}
\title{Check xpdb}
\usage{
check_xpdb(xpdb, check = "data")
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{check}{The `xpdb` slot to be checked, can be `data`, `files`, 
`summary`, `special`, `code`, a file name, or `FALSE`. IF `FALSE` the slot 
test will be skipped.}
}
\value{
Silent if successful check, else returns error.
}
\description{
Check xpdb
}
\keyword{internal}
