% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_edits.R
\name{filter.xpose_data}
\alias{filter.xpose_data}
\title{Filter an xpdb by matching condition}
\usage{
\method{filter}{xpose_data}(.data, ..., .problem, .source)
}
\arguments{
\item{.data}{An xpose database object.}

\item{...}{Logical predicates defined in terms of the variables in .data. 
Multiple conditions are combined with &. Only rows where the condition evaluates 
to \code{TRUE} are kept. 
These arguments are automatically quoted and evaluated in the 
context of the data frame. They support unquoting and splicing. 
See the dplyr vignette("programming") for an introduction to these concepts.}

\item{.problem}{The problem from which the data will be modified}

\item{.source}{The source of the data in the xpdb. Can either be 'data' or an output 
file extension e.g. 'phi'.}
}
\description{
Use filter() find rows/cases where conditions are true. 
Unlike base subsetting, rows where the condition evaluates to NA are dropped.
}
\examples{
xpdb_ex_pk \%>\% 
 filter(DV < 1, .problem = 1) \%>\% 
  dv_vs_ipred()
}
