% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_access.R
\name{get_code}
\alias{get_code}
\title{Access model code}
\usage{
get_code(xpdb, .problem = NULL)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object from which the model code will be extracted.}

\item{.problem}{The problem to be used, in addition, problem 0 is attributed to 
general output (e.g. NM-TRAN warnings in NONMEM). By default returns the 
entire code.}
}
\value{
A tibble of the parsed model.
}
\description{
Access model code from an xpdb object.
}
\examples{
parsed_model <- get_code(xpdb_ex_pk)
parsed_model

}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_model}}
}
