% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_file}
\alias{get_file}
\title{Extract model output file data}
\usage{
get_file(xpdb, file = NULL, problem = NULL, subprob = NULL)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object from which the model output file data will be extracted.}

\item{file}{Name of the file to be extracted from the xpdb e.g. 'run001.ext'.}

\item{problem}{The problem to be used, by default returns the last one for each file.}

\item{subprob}{The subproblem to be used, by default returns the last one for each file.}
}
\value{
A tibble for single file or a named list for multiple files.
}
\description{
Extract model output file data from an xpdb object.
}
\examples{
# Single file (returns a tibble)
ext_file <- get_file(xpdb_ex_pk, 'run001.ext')
ext_file

# Multiple files (returns a list)
files <- get_file(xpdb_ex_pk, c('run001.ext', 'run001.phi'))
files

# Tip to list available files in the xpdb
print(xpdb_ex_pk)

}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_files}}
}
