% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpdb_access.R
\name{get_prm}
\alias{get_prm}
\title{Access model parameters}
\usage{
get_prm(xpdb, problem = NULL, subprob = NULL, method = NULL, signif = 4,
  show_all = FALSE, quiet)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object from which the model output file data will be extracted.}

\item{problem}{The problem to be used, by default returns the last one for each file.}

\item{subprob}{The subproblem to be used, by default returns the last one for each file.}

\item{method}{The estimation method to be used, by default returns the last one for each file}

\item{signif}{The number of significant digits to be displayed.}

\item{show_all}{Logical, whether the 0 fixed off-diagonal elements should be removed outputed or not.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\value{
A tibble for single problem/subprob or a named list for multiple problem|subprob.
}
\description{
Access model parameter estimates from an xpdb object.
}
\examples{
# Display to the console
get_prm(xpdb_ex_pk, problem = 1)

# Store the parameters in an object
prm <- get_prm(xpdb_ex_pk, problem = 1)

}
