% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_get.R
\name{get_var_labels_units}
\alias{get_var_labels_units}
\title{Extract variable labels and units}
\usage{
get_var_labels_units(xpdb, ..., .problem = NULL)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{...}{variable(s) for which labels and units are to be extracted.}

\item{.problem}{The $problem number to be used.}
}
\value{
A character string.
}
\description{
Utility function to construct a string made of a variable label 
and units, if present in xpose object. Units are surrounded by parenthesis or
sqaure brackets based upon 'xpose_square_bracket' option.
}
\examples{
labels_units <- data.frame(
  col   = c('ALAG1', 'CL', 'V'),
  label = c('Lag time', 'Clearance', 'Volume'),
  units = c('h', 'L/h', 'L')
)

xpdb_2 <- set_var_labels_units(xpdb_ex_pk, .problem = 1, info = labels_units)

get_var_labels_units(xpdb_2, 'ALAG1')
get_var_labels_units(xpdb_2, 'ALAG1', 'CL')

}
\keyword{internal}
