% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_tables.R
\name{index_table}
\alias{index_table}
\title{Index table columns}
\usage{
index_table(name, data)
}
\arguments{
\item{name}{the name of the dataset to be indexed}

\item{data}{the data associated with the dataset specified in `name`.}
}
\value{
A tibble of the index.
}
\description{
Index table columns
}
\keyword{internal}
