% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.xpdb}
\alias{is.xpdb}
\title{Test for xpdb class}
\usage{
is.xpdb(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical value, `TRUE` for `xpdb` class 
and `FALSE` otherwise.
}
\description{
Reports whether x is an `xpdb` object
}
\keyword{internal}
