% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_xpose_data.R
\name{print.xpose_data}
\alias{print.xpose_data}
\title{Print an xpose_data object}
\usage{
\method{print}{xpose_data}(x, ...)
}
\arguments{
\item{x}{An \code{xpose_data} object generated with \code{\link{xpose_data}}.}

\item{...}{Ignored in this function}
}
\description{
This function returns to the console a list of the files and options 
attached to an \code{\link{xpose_data}} object.
}
\examples{
# Using the print function
print(xpdb_ex_pk)

# Or simply by writting the xpdb name
xpdb_ex_pk

}
