% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_files.R
\name{read_nm_files}
\alias{read_nm_files}
\title{NONMEM output file import function}
\usage{
read_nm_files(files = NULL, runno = NULL, dir = NULL, prefix = "run",
  ext = c(".ext", ".cor", ".cov", ".phi", ".grd"), quiet = FALSE)
}
\arguments{
\item{files}{Full model file names. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{ext}.}

\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model files.}

\item{prefix}{Prefix of the model file names.}

\item{ext}{A vector of the file extension to import. By default '.ext', '.cor', '.cov', '.phi' and '.grd' 
files are listed.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}
}
\description{
Quickly import NONMEM output files into R.
}
\examples{
\dontrun{
out <- read_nm_files(files = 'inst/extdata/run001.ext')
}
}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_tables}}
}
