% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_model.R
\name{read_nm_model}
\alias{read_nm_model}
\title{NONMEM model file parser}
\usage{
read_nm_model(file = NULL, runno = NULL, dir = NULL, prefix = "run",
  ext = c(".lst", ".mod", ".ctl", ".txt"), verbose = TRUE)
}
\arguments{
\item{file}{Full file name preferably a `.lst` file. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{ext}.}

\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model file.}

\item{prefix}{Prefix of the model file name.}

\item{ext}{Extension of the model file.}

\item{verbose}{Logical, if \code{TRUE} messages are printed to the console.}
}
\value{
A \code{\link[dplyr]{tibble}} of class \code{mod_file} containing the following columns: 
\itemize{
 \item{\strong{problem}}{: a numeric identifier for the $PROBLEM associated with the code.}
 \item{\strong{level}}{: a unique numeric identifier to each subroutine block associated with the code.}
 \item{\strong{subroutine}}{: a character identifier named after the 3 first letters of the subroutine name e.g. "$THETA" and 
 "$TABLE" will become "the" and "tab" respectively. In addtion all output from the .lst is labeled "lst", the general nonmem 
 output e.g. NM-TRAN messages are labeled "oth". With priors thp, tpv, omp, opd, sip, spd abreviations are given to the THETAP, 
 etc.}
 \item{\strong{code}}{: the code without comments or subroutine names e.g. "$THETA 0.5 ; TVCL" will return 0.5.}
 \item{\strong{comment}}{: the last comment of a record e.g. "0.5 ; Clearance (L/h) ; TVCL" will return "TVCL".}
}
}
\description{
Parse NONMEM model files in R format
}
\examples{
\dontrun{
mod_file <- read_nm_model(dir = '../models/pk/', runno = '001')
}
}
\seealso{
\code{\link{xpose_data}}, \code{\link{read_nm_tables}}
}
