% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_tab.R
\name{read_nm_tab}
\alias{read_nm_tab}
\title{NONMEM output table import function}
\usage{
read_nm_tab(file = NULL, skip = NULL, header = NULL,
  rm_duplicates = FALSE, nonmem_tab = TRUE, index = FALSE)
}
\arguments{
\item{file}{Full file name.}

\item{skip}{Number of lines to skip before reading data.}

\item{header}{Logical value indicating whether the file has header or not.}

\item{rm_duplicates}{Logical value indicating whether duplicated columns should be removed.}

\item{nonmem_tab}{Logical value indicading to the function whether the file is a
table or an additional NONMEM output file.}

\item{index}{Logical, when \code{nonmem_tab} is \code{TRUE} returns a list with the data and an
index of all filenames and their colnames.}
}
\description{
Quickly import NONMEM output tables into R.
When both \code{skip} and \code{header} are \code{NULL},
\code{read_nm_tab} will automatically detect the optimal
settings to import the tables. When more than one files are
provided for a same NONMEM run, they will be combined into
a single \code{data.frame}.
}
\examples{
\dontrun{
data <- read_nm_tab(file = '../models/pk/sdtab101')
}
}
