% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_tab.R
\name{read_nm_tab}
\alias{read_nm_tab}
\title{NONMEM output table import function}
\usage{
read_nm_tab(file = NULL, skip = NULL, header = NULL,
  rm_duplicates = FALSE, index = FALSE)
}
\arguments{
\item{file}{Path to the file.}

\item{skip}{Number of lines to skip before reading data.}

\item{header}{Logical value indicating whether the file has a header row or not.}

\item{rm_duplicates}{Logical value indicating whether duplicated columns should be removed.}

\item{index}{Logical value indiacating whether the data should be returned as a simple data.frame or 
a list containing the data and an index of the colum names.}
}
\description{
Quickly import NONMEM output tables into R.
When both \code{skip} and \code{header} are \code{NULL},
\code{read_nm_tab} will automatically detect the optimal
settings to import the tables. When more than one files are
provided for a same NONMEM run, they will be combined into
a single \code{data.frame}.
}
\examples{
\dontrun{
data <- read_nm_tab(file = '../models/pk/sdtab101')
}
}
