% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_tables.R
\name{read_nm_tables}
\alias{read_nm_tables}
\title{NONMEM output table import function}
\usage{
read_nm_tables(file = NULL, dir = NULL, combined = TRUE,
  rm_duplicates = TRUE, quiet = FALSE, simtab = NULL, ziptab = TRUE,
  ...)
}
\arguments{
\item{file}{A character vector of path to the files or a \code{nm_table_list} object created with \code{list_nm_tables}.}

\item{dir}{Location of the model files.}

\item{combined}{Logical value indicating whether multiple tables should be combined into a single one. If the number of rows 
does not match an error will be returned.}

\item{rm_duplicates}{Logical value indicating whether duplicated columns should be removed.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{simtab}{If \code{TRUE} only reads in simulation tables, if \code{FALSE} only reads estimation tables. 
Default \code{NULL} reads all tables.}

\item{ziptab}{If \code{TRUE} search for the tables that have been compressed and renamed ´<file>.zip'.}

\item{...}{Additional arguments to be passed to the \code{\link[readr]{read_table2}} or \code{\link[readr]{read_csv}} functions.}
}
\description{
Quickly import NONMEM output tables into R. This function automatically 
detects the optimal settings to import the tables from nonmem.
}
\note{
An `ID` column must be present in all data tables. This is required for properly combining/merging tables and removing `NA` values. 
If the `ID` column is missing from a table xpose will return the following warning: \code{Unknown variables: `ID`}.
}
\examples{
\dontrun{
# Import tables manually and return them as a list of individual tables
nm_tables <- read_nm_tables(file = c('sdtab001', 'patab001'), 
                            dir = 'models', combined = FALSE)

# Import tables manually and return them as a single merged table
nm_tables <- read_nm_tables(file = c('sdtab001', 'patab001'), 
                            dir = 'models', combined = TRUE)

# Import tables automatically (used internally by xpose_data())
nm_tables <- read_nm_model(file = 'run001.lst', dir = 'models') \%>\% 
              list_nm_tables() \%>\% 
              read_nm_tables()

# Passing arguments to readr via `...` 
# (e.g. import columns as character and only first 10 rows)
nm_tables <- read_nm_tables(file = 'sdtab001', dir = 'models', 
                            col_type = readr::cols(.default = 'c'), 
                            n_max = 10)

}
}
