% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm_tables.R
\name{read_nm_tables}
\alias{read_nm_tables}
\title{NONMEM output table import function}
\usage{
read_nm_tables(files = NULL, combined = TRUE, rm_duplicates = TRUE,
  quiet = FALSE, simtab = NULL, ...)
}
\arguments{
\item{files}{A character vector of path to the files or a \code{nm_table_list} file created with \code{list_nm_tables}.}

\item{combined}{Logical value indicating whether multiple tables should be combined into a single one. If the number of rows 
does not match an error will be returned.}

\item{rm_duplicates}{Logical value indicating whether duplicated columns should be removed.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{simtab}{If \code{TRUE} only reads in simulation tables, if \code{FALSE} only reads estimation tables. 
Default \code{NULL} reads all tables.}

\item{...}{Additional arguments to be passed to the \code{\link[readr]{read_delim}} functions.}
}
\description{
Quickly import NONMEM output tables into R. This function automatically 
detects the optimal settings to import the tables from nonmem.
}
\examples{
\dontrun{
nm_tables <- read_nm_tables(files = c('sdtab001', 'patab001'))
}
}
