% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{facet_wrap_paginate}
\alias{reexports}
\alias{facet_grid_paginate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}}

  \item{ggforce}{\code{\link[ggforce]{facet_wrap_paginate}}, \code{\link[ggforce]{facet_grid_paginate}}}

  \item{purrr}{\code{\link[purrr]{\%>\%}}}
}}

