% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resiudals.R
\name{res_vs_pred}
\alias{res_vs_pred}
\title{Residuals plotted against population predictions}
\usage{
res_vs_pred(xpdb, aes = NULL, res = "CWRES", group = "ID", type = "pls",
  facets = NULL, title = "CWRES vs. PRED | @run", subtitle = "Ofv: @ofv",
  caption = "@dir", log = NULL, guides = TRUE, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{aes}{The xpose aesthetics (eg. \code{point_color}).}

\item{res}{Type of residual to be used. Default is "CWRES".}

\item{group}{Grouping variable to be used for lines.}

\item{type}{String setting the type of plot to be used points 'p',
line 'l' and smooth 's' or any combination of the three.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} or a formula 
to use \link[ggplot2]{facet_grid}.}

\item{title}{Main title of the plot. If NULL automated title will be generated.
Use \code{FALSE} to remove title and subtitle.}

\item{subtitle}{the plot subtitle. If NULL automated subtitle will be generated.
Use \code{FALSE} to remove subtitle.}

\item{caption}{page caption. If NULL automated caption will be generated.
Use \code{FALSE} to remove caption.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 'x', y' or 'xy'.}

\item{guides}{Enable guides display (e.g. unity line).}

\item{...}{Any additional aesthetics to be passed on \code{xplot_scatter}.}
}
\description{
Model residuals plotted against population predictions (PRED). 

The residuals can be one of:
\itemize{
  \item RES: model residuals
  \item WRES: weighted model residuals
  \item CWRES: conditional weighted model residuals
  \item EWRES/ECWRES: Monte Carlo based model residuals
  \item NPDE: Normalized prediction distribution error 
}
}
\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
res_vs_pred(xpdb_ex_pk, res = 'CWRES')

}
\seealso{
\code{\link{xplot_scatter}}
}
