% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_readable.R
\name{theme_readable}
\alias{theme_readable}
\title{ggplot2 theme}
\usage{
theme_readable(base_size = 11, base_family = "",
  legend_position = "right")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{legend_position}{the position of legends ("none", "left", "right", "bottom",
"top", or two-element numeric vector)}
}
\description{
Theme intended to make ggplot2 more readable when used
in presentation or papers. Background and grid lines were dimed to bring the 
focus on the data.
}
\examples{
\dontrun{
ggplot(diamonds, aes(x = price, y = carat, color = cut)) +
 geom_point(alpha = 0.5) +
 theme_readable(legend_position = 'top')
}
}
