% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc.R
\name{vpc}
\alias{vpc}
\title{Visual predictive checks}
\usage{
vpc(xpdb, vpc_opt = NULL, mapping = NULL, type = "alp", smooth = TRUE,
  facets = NULL, title = "Visual predictive checks | @run",
  subtitle = "Nsim: @nsim, PI: @vpcpi", caption = "@dir", log = NULL,
  psn_folder = NULL, obs_problem, sim_problem, quiet, ...)
}
\arguments{
\item{xpdb}{An xpose database object.}

\item{vpc_opt}{A list of options regarding binning, pi and ci computation. 
For more information see \code{\link{vpc_opt_set}}.}

\item{mapping}{List of aesthetics mappings to be used for the xpose plot 
(e.g. \code{point_color}).}

\item{type}{String setting the type of plot to be used points 'p',
line 'l', area 'a' and text 't' or any combination of the four.}

\item{smooth}{Should the bins be smoothed (connect bin midpoints, default) or shown as rectangular boxes.}

\item{facets}{Either a character string to use \link[ggplot2]{facet_wrap} 
or a formula to use \link[ggplot2]{facet_grid}.}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{log}{String assigning logarithmic scale to axes, can be either '', 
'x', y' or 'xy'.}

\item{psn_folder}{Specify a PsN-generated VPC-folder.}

\item{obs_problem}{Alternative to the option `psn_folder`. The $problem number to be used for observations.
By default returns the last estimation problem.}

\item{sim_problem}{Alternative to the option `psn_folder`. The $problem number to be used for simulations.
By default returns the last simulation problem.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{any additional aesthetics.}
}
\description{
Generate visual predictive checks (VPC)
}
\section{Layers mapping}{

Plots can be customized by mapping arguments to specific layers. The naming convention is 
layer_option where layer is one of the names defined in the list below and option is 
any option supported by this layer e.g. point_color = 'blue', smooth_method = 'lm', etc.
\itemize{
  \item point: options to \code{geom_point}
  \item line: options to \code{geom_line}
  \item guide: options to \code{geom_abline}
  \item panel: options to \code{facet_wrap} (facets is character) or \code{facet_grid} 
  (facets is a formula)
  \item smooth: options to \code{geom_smooth}
  \item text: options to \code{geom_text}
  \item xscale: options to \code{scale_x_continuous} or \code{scale_x_log10}
  \item yscale: options to \code{scale_y_continuous} or \code{scale_y_log10}
}
}

\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
\dontrun{
# Simple vpc using simulated data from the xpdb
vpc(xpdb_ex_pk)
}
}
\seealso{
\code{\link[vpc]{vpc}}
}
