% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplot_scatter.R
\name{xplot_scatter}
\alias{xplot_scatter}
\title{Default xpose scatter plot function}
\usage{
xplot_scatter(xpdb, vars = NULL, aes = NULL, group = "ID", type = "pls",
  guides = TRUE, xscale = "continuous", yscale = "continuous",
  title = NULL, subtitle = NULL, caption = NULL,
  plot_name = "scatter_plot", gg_layers, gg_theme, xp_theme, quiet, problem,
  ...)
}
\arguments{
\item{xpdb}{An \code{xpose_data} object generated with \code{\link{xpose_data}}.}

\item{vars}{Variable mapping using the \code{\link[ggplot2]{aes}} function.}

\item{aes}{xpose aesthetics (eg. \code{point_color}).}

\item{group}{Grouping variable to be used for lines.}

\item{type}{String setting the type of plot to be used points 'p',
line 'l', smooth 's' and text 't' or any combination of the 4.}

\item{guides}{should the guides (eg. unity line) be displayed.}

\item{xscale}{scale type for x axis (eg. 'continuous', 'discrete', 'log10').}

\item{yscale}{scale type for y axis (eg. 'continuous', 'discrete', 'log10').}

\item{title}{Plot title. Use \code{NULL} to remove.}

\item{subtitle}{Plot subtitle. Use \code{NULL} to remove.}

\item{caption}{Page caption. Use \code{NULL} to remove.}

\item{plot_name}{name that will be used by \code{xpose_save()} to save the plot.}

\item{gg_layers}{A list of additional ggplot2 layers to be added to the plot.}

\item{gg_theme}{A ggplot2 theme object (eg. \code{\link[ggplot2]{theme_classic}}).}

\item{xp_theme}{An xpose theme or vector of modifications to the xpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')}).}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{problem}{Numeric, the $problem number to use for ploting. By default the data 
is taken from the estimation problem.}

\item{...}{any additional aesthetics.}
}
\description{
Manually generate scatter plots from an xpdb object.
}
\section{Template titles}{

Template titles can be used to create highly informative diagnostics plots. 
They can be applied to any plot title, subtitle and caption. Template titles 
are defined via a single string containing key variables staring with a @ (e.g. @ofv)
which will be replaced by their actual value when rendering the plot.
For example '@run, @nobs observations in @nind subjects' would become 'run001, 
1022 observations in 74 subjects'. The available key variables are listed under 
\code{\link{template_titles}}.
}

\examples{
xplot_scatter(xpdb_ex_pk, aes(x = IPRED, y = DV))

}
