% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose.R
\docType{package}
\name{xpose-package}
\alias{xpose}
\alias{xpose-package}
\title{xpose: graphical diagnostics for pharmacometric models}
\description{
xpose was designed as a ggplot2-based alternative to xpose4. This package aims to reduce the post 
processing burden and improve diagnostics commonly associated the development of non-linear mixed effect models.

To learn more about xpose, visit our website: \url{https://UUPharmacometrics.github.io/xpose/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UUPharmacometrics/xpose}
  \item Report bugs at \url{https://github.com/UUPharmacometrics/xpose/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Guiastrennec \email{guiastrennec@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Andrew C. Hooker \email{andrew.hooker@farmbio.uu.se} [copyright holder]
  \item Sebastian Ueckert \email{sebastian.ueckert@farmbio.uu.se} [copyright holder]
  \item Mats O. Karlsson \email{mats.karlsson@farmbio.uu.se} [copyright holder]
}

Other contributors:
\itemize{
  \item Mike K. Smith \email{mike.k.smith@pfizer.com} [contributor]
}

}
