% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_data.R
\name{xpose_data}
\alias{xpose_data}
\title{Import NONMEM output into R}
\usage{
xpose_data(runno = NULL, dir = NULL, file = NULL, prefix = "run",
  ext = ".lst", gg_theme = theme_readable(),
  xp_theme = theme_xp_default(), simtab = NULL, manual_import = NULL,
  skip = NULL, extra_files, quiet, ...)
}
\arguments{
\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model files.}

\item{file}{Full file name preferably a `.lst` file. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{ext}.}

\item{prefix}{Prefix of the model file name.}

\item{ext}{Extension of the model file. Should be one of ".lst" (default), ".out", ".res", ".mod" or ".ctl" for NONMEM.}

\item{gg_theme}{A ggplot2 theme object (eg. \code{\link[ggplot2]{theme_classic}}).}

\item{xp_theme}{An xpose theme or vector of modifications to the xpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')}).}

\item{simtab}{If \code{TRUE} only reads in simulation tables, if \code{FALSE} only reads estimation tables. 
Default \code{NULL} reads all tables. Option not compatible with manual_import.}

\item{manual_import}{If \code{NULL} (default) the names of the output tables to import will be obtained from the model file. 
To manually import files as in previous versions of xpose, the check the function \code{\link{manual_nm_import}}.}

\item{skip}{Character vector be used to skip the import/generation of: 'data', 'files', 'summary' or any
combination of the three.}

\item{extra_files}{A vector of additional output file extensions to be imported. Default is ".ext", ".cov", ".cor", ".phi", 
".grd" for NONMEM.}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Additional arguments to be passed to the \code{\link[readr]{read_delim}} functions.}
}
\description{
Gather model outputs into a R database
}
\examples{
\dontrun{
xpdb <- xpose_data(file = 'run001.lst')
}

}
