% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_data.R
\name{xpose_data}
\alias{xpose_data}
\title{Import NONMEM output into R}
\usage{
xpose_data(runno = NULL, dir = NULL, prefix = "run", ext = ".lst",
  file = NULL, rounding = NULL, gg_theme = theme_readable(),
  xp_theme = theme_xp_default(), verbose = TRUE)
}
\arguments{
\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model files.}

\item{prefix}{Prefix of the model file name.}

\item{ext}{Model file extention.}

\item{file}{Full file name as an alternative to \code{dir}, \code{prefix}.}

\item{rounding}{Number of significant digits to be used on model info numerical
values (eg. OFV, shrinkage).}

\item{gg_theme}{A ggplot2 complete theme object (eg. \code{theme_classic()}).}

\item{xp_theme}{A xpose theme or vector of modifications of the xpose theme#' \code{runno} and \code{ext}.}

\item{verbose}{Logical, if \code{TRUE} messages are printed to the console.}
}
\description{
Import NONMEM output into a R database
}
\examples{
\dontrun{
xpdb <- xpose_data(dir = '../models/pk/', runno = '001')
}
}
