% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_data.R
\name{xpose_data}
\alias{xpose_data}
\title{Import NONMEM output into R}
\usage{
xpose_data(runno = NULL, dir = NULL, file = NULL, prefix = "run",
  ext = ".lst", gg_theme = theme_readable(),
  xp_theme = theme_xp_default(), quiet = FALSE, extra_files,
  manual_import = TRUE, table_names, tab_suffix, sim_suffix, ...)
}
\arguments{
\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model files.}

\item{file}{Full file name preferably a `.lst` file. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{ext}.}

\item{prefix}{Prefix of the model file name.}

\item{ext}{Extension of the model file. Should be one of ".lst" (default), ".out", ".res", ".mod" or ".ctl" for NONMEM.}

\item{gg_theme}{A ggplot2 theme object (eg. \code{\link[ggplot2]{theme_classic}}).}

\item{xp_theme}{An xpose theme or vector of modifications to the xpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')}).}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{extra_files}{A vector of additional output file extensions to be imported. Default is ".ext", ".cov", ".cor", ".phi", ".grd" for NONMEM.}

\item{manual_import}{Logical, if \code{FALSE} the names of the output tables to import will be collected from the model file (default). If \code{TRUE} the
names of the output tables to import will be generated by combining the table_names argument with the tab_suffix or sim_suffix arguments.}

\item{table_names}{only used when \code{manual_import} is set to \code{TRUE}}

\item{tab_suffix}{only used when \code{manual_import} is set to \code{TRUE}}

\item{sim_suffix}{only used when \code{manual_import} is set to \code{TRUE}}

\item{...}{Additional arguments to be passed to the \code{\link[readr]{read_delim}} functions.}
}
\description{
Gather model outputs into a R database
}
\examples{
\dontrun{
xpdb <- xpose_data(file = 'run001.lst')
}

}
