% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_data.R
\name{xpose_data}
\alias{xpose_data}
\title{Import NONMEM output into R}
\usage{
xpose_data(runno = NULL, dir = NULL, prefix = "run", ext = ".lst",
  file = NULL, rounding = NULL, gg_theme = theme_readable(),
  xp_theme = xp_theme_default(), verbose = FALSE)
}
\arguments{
\item{runno}{run number to be evaluated}

\item{dir}{location of the model files}

\item{prefix}{prefix of the model file name}

\item{ext}{model file extention}

\item{file}{full file name as an alternative to \code{dir}, \code{prefix},}

\item{rounding}{number of significant digits to be used on model info numerical
values (eg. OFV, shrinkage).}

\item{gg_theme}{a ggplot2 complete theme object (eg. \code{theme_classic()})}

\item{xp_theme}{a ggxpose theme or vector of modifications of the ggxpose theme#' \code{runno} and \code{ext}}

\item{verbose}{if \code{TRUE} messages are printed to the console}
}
\description{
Import NONMEM output into a R database
}
\examples{
\dontrun{
xpdb <- xpose_data(dir = '../models/pk/', runno = '001')
}
}
