% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_data.R
\name{xpose_data}
\alias{xpose_data}
\title{Import NONMEM output into R}
\usage{
xpose_data(file = NULL, runno = NULL, dir = NULL, prefix = "run",
  ext = ".lst", gg_theme = theme_readable(),
  xp_theme = theme_xp_default(), quiet = FALSE, ...)
}
\arguments{
\item{file}{Full file name preferably a `.lst` file. Alternative argument to \code{dir}, \code{prefix},
\code{runno} and \code{ext}.}

\item{runno}{Run number to be evaluated.}

\item{dir}{Location of the model files.}

\item{prefix}{Prefix of the model file name.}

\item{ext}{Extension of the model file.}

\item{gg_theme}{A ggplot2 theme object (eg. \code{\link[ggplot2]{theme_classic}}).}

\item{xp_theme}{An xpose theme or vector of modifications to the xpose theme
(eg. \code{c(point_color = 'red', line_linetype = 'dashed')}).}

\item{quiet}{Logical, if \code{FALSE} messages are printed to the console.}

\item{...}{Additional arguments to be passed to the \code{\link[readr]{read_delim}} functions.}
}
\description{
Gather model outputs into a R database
}
\examples{
\dontrun{
xpdb <- xpose_data(file = 'run001.lst')
}

}
