% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_labs.R
\name{xpose_labs}
\alias{xpose_labs}
\title{Modify labels using template titles}
\usage{
xpose_labs(...)
}
\arguments{
\item{...}{A list of new name-value pairs. The name should either be an aesthetic, 
or one of "title", "subtitle", or "caption". The value should be a text containing 
\code{\link{template_titles}} keywords for the axes, plot title, subtitle or caption.}
}
\description{
Function based on \link[ggplot2]{labs} enabling the 
use of \code{\link{template_titles}} using the ggplot2 \code{+} operator.
}
\examples{
\dontrun{
dv_vs_ipred(xpdb_ex_pk) + 
  xpose_labs(title = 'DV vs. IPRED | @run', 
             caption = '@dir', 
             x = 'Individual Predictions (IPRED)')
}
}
\seealso{
\link[ggplot2]{ggplot2}, \link[ggplot2]{labs}
}
