% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_save.R
\name{xpose_save}
\alias{xpose_save}
\title{Save ggxpose plot}
\usage{
xpose_save(plot = last_plot(), filename = "@run_@plotfun.pdf", dir = NULL,
  width = 8, height = 7, res = 200, ...)
}
\arguments{
\item{plot}{a ggxpose plot object.}

\item{filename}{an optional name to be given to the file. Template variables such as @run 
can be used to generate template names. By default, the file extension will be set to .pdf 
by default but can be changed to .jpeg, .png, .bmp, .tiff}

\item{dir}{an optional path to a specific directory.}

\item{width}{the page width in inches}

\item{height}{the page height in inches}

\item{res}{the nominal resolution in ppi. Not used with .pdf}

\item{...}{additional options to be passed to ploting functions.}
}
\description{
Automatically save ggxpose plots to pdf, jpg or png.

If no arguments are provided, the function will automatically name
and save a plot after its run number and the plotting function name
by using the metadata attached to the plot.
}
\examples{
\dontrun{
xpdb \%>\% 
 dv_vs_ipred() \%>\% 
 xpose_save()
}
}
