% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_save.R
\name{xpose_save}
\alias{xpose_save}
\title{Save xpose plot}
\usage{
xpose_save(plot = last_plot(), file = "@run_@plotfun.pdf", dir = NULL,
  device = NULL, width = 7, height = 6, units = c("in", "cm", "mm"),
  dpi = 200, ...)
}
\arguments{
\item{plot}{A xpose plot object.}

\item{file}{An optional name to be given to the file. Template variables such as @run 
can be used to generate template names.}

\item{dir}{Directory under which the xpose plot will be saved. Template variables such as @dir 
can be used to generate template names.}

\item{device}{Graphical device to use. Can be either be a device function
(e.g. \code{\link{png}}), or one of 'eps', 'ps', 'tex' (pictex),
'pdf' (default), 'jpeg', 'tiff', 'png', 'bmp', 'svg' or 'wmf' (windows only).}

\item{width, height}{Plot size in \code{units}.}

\item{units}{Units of the plot size ('in', 'cm', or 'mm').}

\item{dpi}{Plot resolution. Applies only to raster output types.}

\item{...}{Additional arguments passed on to \link[ggplot2]{ggsave} or graphics \code{device}.}
}
\description{
Automatically save xpose plots to files.

This function was inspired by the \link[ggplot2]{ggsave}. If no arguments are provided, 
the function will automatically name and save a plot after its run number and the plotting 
function name by using the metadata attached to the plot.
}
\examples{
\dontrun{
xpdb_ex_pk \%>\% 
 dv_vs_ipred() \%>\% 
 xpose_save()
}
}
